Glance at a mixpoissonreg object

Glance accepts a mixpoissonreg object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are Efron's pseudo-\(R^2\), degrees of freedom, AIC, BIC, log-likelihood, the type of model used in the fit ('NB' or 'PIG'), the total number of observations and the estimation method.

# S3 method for mixpoissonreg
glance(x, ...)

Arguments

x
A mixpoissonreg object.

...
Additional arguments. Currently not used.

Value

A tibble::tibble() with exactly one row and columns:

efron.pseudo.r2 Efron's pseudo-\(R^2\), that is, the squared correlation between the fitted values and the response values.

df.null Degrees of freedom used by the null model.

logLik The log-likelihood of the model.

AIC Akaike's Information Criterion for the model.

BIC Bayesian Information Criterion for the model.

df.residual Residual degrees of freedom.

nobs Number of observations used.

model.type Type of model fitted, "NB" or "PIG".

est.method The estimation method of the fitted model, "EM" or "ML".

See also