Predict Method for mixpoissonreg Objects

Function to obtain various predictions based on the fitted mixed Poisson regression models.

# S3 method for mixpoissonreg
predict(
  object,
  newdata = NULL,
  type = c("response", "link", "precision", "variance"),
  se.fit = FALSE,
  interval = c("none", "confidence", "prediction"),
  level = 0.95,
  nsim_pred = 100,
  nsim_pred_y = 100,
  ...
)

Arguments

object
object of class "mixpoissonreg" containing results from the fitted model.

newdata
optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted response values will be provided.

type
the type of prediction. The default is the "response" type, which provided the estimated values for the means. The type "link" provides the estimates for the linear predictor. The type "precision" provides estimates for the precision parameters whereas the type "variance" provides estimates for the variances.

se.fit
logical switch indicating if standard errors on the scale of linear predictors should be returned. If TRUE, it only returns the standard deviations of the linear predictors when type = 'link', otherwise returns NA and a warning indicating that the type must be 'link'. When using mixpoissonreg objects, the fit must be done with x = TRUE.

interval
Type of interval calculation for the response variables, 'none', 'confidence' or 'prediction'. If 'confidence', the confidence intervals for the means are returned. If 'prediction', prediction intervals for future response variables are reported. For confidence intervals, the type of the prediction must be 'response' or 'link'. For prediction intervals the type of prediction must be 'response'. For 'confidence' intervals, when using mixpoissonreg objects, the fit must be done with x = TRUE and for predictions intervals, the fit must be done with x = TRUE and w = TRUE.

level
Tolerance/confidence level. The default is set to 0.95.

nsim_pred
number of means and predictions to be generated in each step of the simulation. The default is set to 100.

nsim_pred_y
number of response variables generated for each pair of mean and precision to compute the prediction intervals. The default is set to 100.

...
further arguments passed to or from other methods.

Details

The se.fit argument only returns a non-NA vector for type = 'link', that is, on the scale of the linear predictor for the mean parameter. For the response scale, one can obtain confidence or prediction intervals. It is important to notice that confidence intervals must not be used for future observations as they will underestimate the uncertainty. In this case prediction intervals should be used. Currently, we do not have closed-form expressions for the prediction interval and, therefore, they are obtained by simulation and can be computationally-intensive.

See also

Examples

# \donttest{
data("Attendance", package = "mixpoissonreg")

daysabs_fit <- mixpoissonreg(daysabs ~ gender + math +
prog | gender + math + prog, data = Attendance)
predict(daysabs_fit, interval = "confidence")

#>           fit       lwr       upr
#> 1    5.249744  4.139949  6.657040
#> 2    6.661848  5.400526  8.217759
#> 3    8.915212  7.211905 11.020805
#> 4    9.154336  7.343178 11.412207
#> 5   10.042920  7.759159 12.998862
#> 6    6.361626  5.017382  8.066015
#> 7    6.707466  5.394408  8.340136
#> 8    7.808468  6.044819 10.086682
#> 9    5.683602  4.582766  7.048871
#> 10   2.477117  1.856700  3.304846
#> 11   3.565293  2.608368  4.873282
#> 12   6.885947  5.549313  8.544531
#> 13   6.277988  4.924336  8.003747
#> 14   9.361068  7.124103 12.300440
#> 15   3.206450  2.244510  4.580654
#> 16   4.419971  3.228946  6.050316
#> 17   8.126406  6.697559  9.860081
#> 18   8.072810  6.657360  9.789205
#> 19   7.326643  5.397694  9.944932
#> 20   7.756968  6.021121  9.993249
#> 21   8.856413  7.178072 10.927177
#> 22   5.319682  4.213768  6.715847
#> 23   6.072417  4.941114  7.462740
#> 24   5.249744  4.139949  6.657040
#> 25   8.974401  7.245324 11.116118
#> 26   8.474782  6.531082 10.996942
#> 27   7.654987  5.973035  9.810559
#> 28   8.856413  7.178072 10.927177
#> 29   5.012121  3.884048  6.467829
#> 30   7.230319  5.298228  9.866980
#> 31   7.912494  6.091558 10.277759
#> 32   8.399772  6.891583 10.238022
#> 33   6.112732  4.975818  7.509419
#> 34   6.153316  5.010265  7.557143
#> 35   5.078894  3.956643  6.519455
#> 36   8.915212  7.211905 11.020805
#> 37   7.455028  5.873916  9.461735
#> 38   4.568655  3.394168  6.149552
#> 39  12.114723  8.974893 16.353008
#> 40   8.072810  6.657360  9.789205
#> 41   8.234669  6.776569 10.006504
#> 42   8.126406  6.697559  9.860081
#> 43   7.810085  6.449656  9.457470
#> 44   6.574264  5.338675  8.095819
#> 45   6.977684  5.606512  8.684200
#> 46   8.511676  6.965926 10.400431
#> 47   4.881197  3.740521  6.369723
#> 48   6.194168  5.044446  7.605933
#> 49   6.931664  5.578075  8.613717
#> 50   5.954292  4.560081  7.774774
#> 51   7.164840  5.717075  8.979230
#> 52   5.390552  4.287725  6.777034
#> 53   5.797557  4.692051  7.163534
#> 54   7.072107  5.772673  8.664045
#> 55   9.910884  7.703244 12.751202
#> 56   5.462367  4.361715  6.840761
#> 57  11.798269  8.826254 15.771034
#> 58  11.955449  8.901149 16.057788
#> 59   8.739977  7.109127 10.744948
#> 60   8.072810  6.657360  9.789205
#> 61   6.402535  5.211026  7.866484
#> 62   8.511676  6.965926 10.400431
#> 63   7.308494  5.796786  9.214429
#> 64   5.874794  4.764177  7.244315
#> 65   8.344373  6.853713 10.159247
#> 66   6.445042  5.243423  7.922032
#> 67  13.118645 10.417421 16.520291
#> 68   9.276293  7.406508 11.618108
#> 69   3.565293  2.608368  4.873282
#> 70   7.407393  6.096677  8.999898
#> 71   4.946226  3.811984  6.417957
#> 72   4.722341  3.564884  6.255605
#> 73   3.102098  2.200061  4.373975
#> 74   6.795417  5.490798  8.410015
#> 75   3.041754  2.290224  4.039897
#> 76  14.297112 11.082570 18.444044
#> 77   6.360308  5.178314  7.812103
#> 78   5.112613  3.993118  6.545966
#> 79   7.912494  6.091558 10.277759
#> 80   8.856413  7.178072 10.927177
#> 81   8.180358  6.737297  9.932509
#> 82   4.753693  3.599640  6.277739
#> 83   8.937351  6.667033 11.980779
#> 84   5.390552  4.287725  6.777034
#> 85   4.849004  3.705035  6.346186
#> 86  10.542990  8.137147 13.660148
#> 87   7.358539  6.051084  8.948496
#> 88   9.337879  7.437630 11.723625
#> 89   1.965005  1.439203  2.682904
#> 90   6.661848  5.400526  8.217759
#> 91   8.511676  6.965926 10.400431
#> 92   6.403861  5.064138  8.098009
#> 93   4.219907  3.007574  5.920922
#> 94   9.804873  7.594224 12.659033
#> 95   9.525102  7.528945 12.050501
#> 96   7.260292  5.770511  9.134692
#> 97   6.977684  5.606512  8.684200
#> 98   7.405859  5.848482  9.377946
#> 99   6.032368  4.906166  7.417087
#> 100  7.860309  6.068296 10.181517
#> 101 10.404379  8.049375 13.448387
#> 102  6.235292  5.078350  7.655805
#> 103  7.308494  5.796786  9.214429
#> 104  7.808468  6.044819 10.086682
#> 105  7.673996  5.752286 10.237707
#> 106  8.511676  6.965926 10.400431
#> 107 15.478678 11.627218 20.605915
#> 108  6.796825  5.489231  8.415900
#> 109  5.953059  4.835585  7.328775
#> 110  6.318360  5.145292  7.758874
#> 111  4.753693  3.599640  6.277739
#> 112  9.976683  7.731327 12.874143
#> 113  6.931664  5.578075  8.613717
#> 114  6.706077  5.430951  8.280588
#> 115  5.953059  4.835585  7.328775
#> 116  7.456572  6.141986  9.052522
#> 117  8.072810  6.657360  9.789205
#> 118 14.778055 11.316534 19.298393
#> 119  5.722521  4.297282  7.620455
#> 120  7.808468  6.044819 10.086682
#> 121  7.654987  5.973035  9.810559
#> 122  6.767363  4.816255  9.508883
#> 123  7.308494  5.796786  9.214429
#> 124  4.568655  3.394168  6.149552
#> 125  4.390820  3.196598  6.031193
#> 126  6.361626  5.017382  8.066015
#> 127 14.109146 10.985826 18.120439
#> 128  7.758575  6.406833  9.395514
#> 129  6.707466  5.394408  8.340136
#> 130  8.126406  6.697559  9.860081
#> 131  7.555910  6.231672  9.161551
#> 132  8.399772  6.891583 10.238022
#> 133  9.976683  7.731327 12.874143
#> 134  6.318360  5.145292  7.758874
#> 135  9.867925  7.683876 12.672763
#> 136  6.885947  5.549313  8.544531
#> 137  5.954292  4.560081  7.774774
#> 138  6.277988  4.924336  8.003747
#> 139  5.045397  3.920284  6.493415
#> 140 13.032123 10.362494 16.389513
#> 141  9.337879  7.437630 11.723625
#> 142  6.154590  4.786118  7.914342
#> 143 14.876168 11.362058 19.477139
#> 144  5.390552  4.287725  6.777034
#> 145  8.399772  6.891583 10.238022
#> 146  6.979129  5.678526  8.577620
#> 147  7.606075  6.276013  9.218013
#> 148  7.166324  5.866261  8.754502
#> 149  8.739977  7.109127 10.744948
#> 150 12.860788 10.250948 16.135081
#> 151  6.112732  4.975818  7.509419
#> 152  5.759320  4.655754  7.124468
#> 153 12.117232  9.721261 15.103730
#> 154  9.933439  7.730795 12.763656
#> 155 13.559946 10.683736 17.210470
#> 156 11.566365  8.711653 15.356536
#> 157  6.235292  5.078350  7.655805
#> 158  5.608879  4.509352  6.976506
#> 159  3.565293  2.608368  4.873282
#> 160  2.349395  1.764924  3.127420
#> 161  2.349395  1.764924  3.127420
#> 162  2.318508  1.741445  3.086791
#> 163  3.612790  2.632629  4.957878
#> 164  7.758575  6.406833  9.395514
#> 165  2.141524  1.597420  2.870959
#> 166  2.962299  2.232546  3.930586
#> 167  2.141524  1.597420  2.870959
#> 168  2.754337  2.068875  3.666909
#> 169  2.058165  1.524327  2.778959
#> 170  6.403861  5.064138  8.098009
#> 171  2.461289  1.809435  3.347975
#> 172  2.318508  1.741445  3.086791
#> 173  2.228259  1.669988  2.973158
#> 174  8.455539  6.928986 10.318413
#> 175  2.058165  1.524327  2.778959
#> 176  2.228259  1.669988  2.973158
#> 177  2.700199  2.023323  3.603514
#> 178  4.136961  2.916384  5.868379
#> 179  3.001764  2.261530  3.984288
#> 180  2.629666  1.962265  3.524062
#> 181  6.979129  5.678526  8.577620
#> 182 11.416665  9.150172 14.244568
#> 183  3.734322  2.691870  5.180473
#> 184  1.965005  1.439203  2.682904
#> 185  2.700199  2.023323  3.603514
#> 186  2.004403  1.475610  2.722692
#> 187 11.045116  8.818996 13.833161
#> 188  2.664699  1.992828  3.563088
#> 189  5.180725  4.066368  6.600462
#> 190 11.645567  9.344566 14.513165
#> 191  2.560445  1.912071  3.428680
#> 192  2.595093  1.931662  3.486381
#> 193  3.834484  2.737885  5.370302
#> 194  9.154336  7.343178 11.412207
#> 195  5.285692  3.803801  7.344900
#> 196  9.740207  7.526464 12.605072
#> 197  2.526783  1.890119  3.377898
#> 198  6.979129  5.678526  8.577620
#> 199  6.532257  5.205169  8.197695
#> 200  2.461289  1.809435  3.347975
#> 201  2.754337  2.068875  3.666909
#> 202  3.809194  2.726491  5.321845
#> 203  3.565293  2.608368  4.873282
#> 204  2.257945  1.693958  3.009705
#> 205  7.606075  6.276013  9.218013
#> 206  2.735605  2.017722  3.708902
#> 207  9.361068  7.124103 12.300440
#> 208  2.349882  1.703748  3.241057
#> 209 11.879059  9.535076 14.799257
#> 210  2.809561  2.114113  3.733782
#> 211  2.923353  2.203292  3.878739
#> 212  1.991183  1.463457  2.709208
#> 213 10.269718  8.071668 13.066335
#> 214  2.257945  1.693958  3.009705
#> 215  2.629666  1.962265  3.524062
#> 216  2.526783  1.890119  3.377898
#> 217  3.249167  2.261937  4.667278
#> 218  6.796825  5.489231  8.415900
#> 219  6.361626  5.017382  8.066015
#> 220  2.846991  2.144052  3.780393
#> 221  7.758575  6.406833  9.395514
#> 222  7.164840  5.717075  8.979230
#> 223  3.784072  2.715025  5.274058
#> 224  5.462367  4.361715  6.840761
#> 225  7.261795  5.959121  8.849237
#> 226  4.881197  3.740521  6.369723
#> 227  2.611782  1.944492  3.508065
#> 228  1.926381  1.402989  2.645027
#> 229  2.198964  1.645894  2.937882
#> 230  5.285692  3.803801  7.344900
#> 231  6.707466  5.394408  8.340136
#> 232  2.846991  2.144052  3.780393
#> 233  2.736172  2.053721  3.645401
#> 234  2.058165  1.524327  2.778959
#> 235 10.001459  7.798337 12.826989
#> 236  6.033617  4.649789  7.829287
#> 237  3.834484  2.737885  5.370302
#> 238  6.995012  5.053847  9.681772
#> 239  6.617911  5.369767  8.156173
#> 240  3.834484  2.737885  5.370302
#> 241  2.884322  2.097299  3.966681
#> 242  6.195451  4.831995  7.943637
#> 243  2.884920  2.173786  3.828693
#> 244  2.646576  1.965762  3.563182
#> 245  6.445042  5.243423  7.922032
#> 246  7.212408  5.743942  9.056294
#> 247  3.041754  2.290224  4.039897
#> 248  8.234669  6.776569 10.006504
#> 249  6.360308  5.178314  7.812103
#> 250  2.809561  2.114113  3.733782
#> 251  3.337008  2.481919  4.486697
#> 252  2.611782  1.944492  3.508065
#> 253  8.399772  6.891583 10.238022
#> 254  2.577444  1.922946  3.454709
#> 255  6.113999  4.740452  7.885531
#> 256  3.082277  2.318612  4.097464
#> 257  3.271416  2.442267  4.382062
#> 258  2.444550  1.834104  3.258170
#> 259  2.198964  1.645894  2.937882
#> 260  2.477117  1.856700  3.304846
#> 261  7.966676  6.575599  9.652037
#> 262  2.228259  1.669988  2.973158
#> 263  9.175169  7.147220 11.778528
#> 264  7.407393  6.096677  8.999898
#> 265  8.092858  5.748904 11.392491
#> 266  3.495214  2.571345  4.751023
#> 267  1.926381  1.402989  2.645027
#> 268  2.085585  1.548712  2.808568
#> 269  6.979129  5.678526  8.577620
#> 270  2.288026  1.717784  3.047568
#> 271  7.523158  5.599202 10.108210
#> 272  2.754337  2.068875  3.666909
#> 273  2.273406  1.629597  3.171567
#> 274  2.113369  1.573084  2.839220
#> 275  2.213563  1.657956  2.955364
#> 276  5.462367  4.361715  6.840761
#> 277  2.058165  1.524327  2.778959
#> 278  2.412411  1.811268  3.213067
#> 279  2.113369  1.573084  2.839220
#> 280  1.965005  1.439203  2.682904
#> 281  4.598987  3.427887  6.170180
#> 282  2.629666  1.962265  3.524062
#> 283  2.560445  1.912071  3.428680
#> 284  2.595093  1.931662  3.486381
#> 285  8.587685  6.635044 11.114973
#> 286  3.293136  2.455573  4.416379
#> 287  2.526783  1.890119  3.377898
#> 288  8.072810  6.657360  9.789205
#> 289  2.884920  2.173786  3.828693
#> 290  2.664699  1.992828  3.563088
#> 291  2.510638  1.855174  3.397688
#> 292  6.072417  4.941114  7.462740
#> 293  3.541778  2.596112  4.831916
#> 294  7.357015  5.822775  9.295512
#> 295  2.288026  1.717784  3.047568
#> 296  5.285692  3.803801  7.344900
#> 297  7.827859  5.906721 10.373838
#> 298  3.472162  2.558833  4.711487
#> 299  2.754337  2.068875  3.666909
#> 300  2.461289  1.809435  3.347975
#> 301  2.155742  1.609567  2.887251
#> 302  5.462367  4.361715  6.840761
#> 303  2.560445  1.912071  3.428680
#> 304  3.685226  2.668410  5.089506
#> 305  2.257945  1.693958  3.009705
#> 306  9.651998  7.588214 12.277074
#> 307  2.257945  1.693958  3.009705
#> 308  2.962299  2.232546  3.930586
#> 309  3.228407  2.415390  4.315084
#> 310  6.706077  5.430951  8.280588
#> 311  2.595093  1.931662  3.486381
#> 312  6.887374  5.583991  8.494985
#> 313  2.412911  1.763925  3.300672
#> 314  2.629666  1.962265  3.524062

predict(daysabs_fit, type = "link", se.fit = TRUE)

#> $fit
#>         1         2         3         4         5         6         7         8 
#> 1.6581792 1.8963969 2.1877590 2.2142277 2.3068679 1.8502839 1.9032212 2.0552088 
#>         9        10        11        12        13        14        15        16 
#> 1.7375851 0.9070953 1.2712461 1.9294827 1.8370496 2.2365594 1.1651645 1.4861331 
#>        17        18        19        20        21        22        23        24 
#> 2.0951188 2.0885016 1.9915174 2.0485916 2.1811419 1.6714136 1.8037567 1.6581792 
#>        25        26        27        28        29        30        31        32 
#> 2.1943762 2.1370949 2.0353573 2.1811419 1.6118591 1.9782831 2.0684431 2.1282046 
#>        33        34        35        36        37        38        39        40 
#> 1.8103739 1.8169910 1.6250934 2.1877590 2.0088886 1.5192189 2.4944215 2.0885016 
#>        41        42        43        44        45        46        47        48 
#> 2.1083531 2.0951188 2.0554159 1.8831626 1.9427171 2.1414389 1.5853905 1.8236082 
#>        49        50        51        52        53        54        55        56 
#> 1.9360999 1.7841124 1.9691857 1.6846479 1.7574366 1.9561585 2.2936336 1.6978822 
#>        57        58        59        60        61        62        63        64 
#> 2.4679528 2.4811872 2.1679075 2.0885016 1.8566940 2.1414389 1.9890372 1.7706709 
#>        65        66        67        68        69        70        71        72 
#> 2.1215874 1.8633112 2.5740345 2.2274620 1.2712461 2.0024786 1.5986248 1.5523047 
#>        73        74        75        76        77        78        79        80 
#> 1.1320787 1.9162484 1.1124343 2.6600575 1.8500768 1.6317106 2.0684431 2.1811419 
#>        81        82        83        84        85        86        87        88 
#> 2.1017360 1.5589219 2.1902393 1.6846479 1.5787733 2.3554611 1.9958614 2.2340791 
#>        89        90        91        92        93        94        95        96 
#> 0.6754948 1.8963969 2.1414389 1.8569011 1.4398130 2.2828795 2.2539306 1.9824200 
#>        97        98        99       100       101       102       103       104 
#> 1.9427171 2.0022715 1.7971396 2.0618259 2.3422268 1.8302254 1.9890372 2.0552088 
#>       105       106       107       108       109       110       111       112 
#> 2.0378375 2.1414389 2.7394634 1.9164555 1.7839053 1.8434597 1.5589219 2.3002507 
#>       113       114       115       116       117       118       119       120 
#> 1.9360999 1.9030141 1.7839053 2.0090957 2.0885016 2.6931433 1.7444094 2.0552088 
#>       121       122       123       124       125       126       127       128 
#> 2.0353573 1.9121115 1.9890372 1.5192189 1.4795160 1.8502839 2.6468232 2.0487987 
#>       129       130       131       132       133       134       135       136 
#> 1.9032212 2.0951188 2.0223301 2.1282046 2.3002507 1.8434597 2.2892896 1.9294827 
#>       137       138       139       140       141       142       143       144 
#> 1.7841124 1.8370496 1.6184763 2.5674173 2.2340791 1.8171981 2.6997605 1.6846479 
#>       145       146       147       148       149       150       151       152 
#> 2.1282046 1.9429242 2.0289472 1.9693928 2.1679075 2.5541830 1.8103739 1.7508195 
#>       153       154       155       156       157       158       159       160 
#> 2.4946286 2.2959067 2.6071203 2.4481014 1.8302254 1.7243508 1.2712461 0.8541580 
#>       161       162       163       164       165       166       167       168 
#> 0.8541580 0.8409237 1.2844804 2.0487987 0.7615178 1.0859657 0.7615178 1.0131769 
#>       169       170       171       172       173       174       175       176 
#> 0.7218149 1.8569011 0.9006852 0.8409237 0.8012208 2.1348218 0.7218149 0.8012208 
#>       177       178       179       180       181       182       183       184 
#> 0.9933255 1.4199615 1.0992000 0.9668568 1.9429242 2.4350741 1.3175662 0.6754948 
#>       185       186       187       188       189       190       191       192 
#> 0.9933255 0.6953462 2.4019884 0.9800911 1.6449449 2.4549256 0.9401811 0.9536225 
#>       193       194       195       196       197       198       199       200 
#> 1.3440349 2.2142277 1.6650035 2.2762623 0.9269468 1.9429242 1.8767526 0.9006852 
#>       201       202       203       204       205       206       207       208 
#> 1.0131769 1.3374177 1.2712461 0.8144551 2.0289472 1.0063527 2.2365594 0.8543651 
#>       209       210       211       212       213       214       215       216 
#> 2.4747771 1.0330284 1.0727314 0.6887291 2.3291996 0.8144551 0.9668568 0.9269468 
#>       217       218       219       220       221       222       223       224 
#> 1.1783988 1.9164555 1.8502839 1.0462627 2.0487987 1.9691857 1.3308005 1.6978822 
#>       225       226       227       228       229       230       231       232 
#> 1.9826271 1.5853905 0.9600326 0.6556433 0.7879865 1.6650035 1.9032212 1.0462627 
#>       233       234       235       236       237       238       239       240 
#> 1.0065598 0.7218149 2.3027310 1.7973467 1.3440349 1.9451973 1.8897798 1.3440349 
#>       241       242       243       244       245       246       247       248 
#> 1.0592899 1.8238153 1.0594971 0.9732669 1.8633112 1.9758028 1.1124343 2.1083531 
#>       249       250       251       252       253       254       255       256 
#> 1.8500768 1.0330284 1.2050745 0.9600326 2.1282046 0.9467983 1.8105810 1.1256686 
#>       257       258       259       260       261       262       263       264 
#> 1.1852231 0.8938610 0.7879865 0.9070953 2.0752673 0.8012208 2.2165008 2.0024786 
#>       265       266       267       268       269       270       271       272 
#> 2.0909819 1.2513946 0.6556433 0.7350492 1.9429242 0.8276894 2.0179861 1.0131769 
#>       273       274       275       276       277       278       279       280 
#> 0.8212793 0.7482835 0.7946036 1.6978822 0.7218149 0.8806267 0.7482835 0.6754948 
#>       281       282       283       284       285       286       287       288 
#> 1.5258361 0.9668568 0.9401811 0.9536225 2.1503292 1.1918402 0.9269468 2.0885016 
#>       289       290       291       292       293       294       295       296 
#> 1.0594971 0.9800911 0.9205367 1.8037567 1.2646290 1.9956543 0.8276894 1.6650035 
#>       297       298       299       300       301       302       303       304 
#> 2.0576890 1.2447775 1.0131769 0.9006852 0.7681350 1.6978822 0.9401811 1.3043319 
#>       305       306       307       308       309       310       311       312 
#> 0.8144551 2.2671649 0.8144551 1.0859657 1.1719887 1.9030141 0.9536225 1.9296898 
#>       313       314 
#> 0.8808338 0.9668568 
#> 
#> $se.fit
#>          1          2          3          4          5          6          7 
#> 0.12117352 0.10709409 0.10817844 0.11247957 0.13163198 0.12111222 0.11115427 
#>          8          9         10         11         12         13         14 
#> 0.13061833 0.10983994 0.14709182 0.15945263 0.11010848 0.12391053 0.13932685 
#>         15         16         17         18         19         20         21 
#> 0.18198155 0.16019550 0.09866285 0.09835822 0.15589350 0.12924632 0.10720150 
#>         22         23         24         25         26         27         28 
#> 0.11890849 0.10518861 0.12117352 0.10919584 0.13292201 0.12658499 0.10720150 
#>         29         30         31         32         33         34         35 
#> 0.13009491 0.15863083 0.13344083 0.10097320 0.10499386 0.10484999 0.12739897 
#>         36         37         38         39         40         41         42 
#> 0.10817844 0.12161813 0.15161517 0.15305915 0.09835822 0.09943149 0.09866285 
#>         43         44         45         46         47         48         49 
#> 0.09764929 0.10621887 0.11162871 0.10225104 0.13580122 0.10475720 0.11084698 
#>         50         51         52         53         54         55         56 
#> 0.13611068 0.11516970 0.11678358 0.10794427 0.10358519 0.12856969 0.11480659 
#>         57         58         59         60         61         62         63 
#> 0.14807523 0.15051645 0.10537345 0.09835822 0.10506173 0.10225104 0.11823355 
#>         64         65         66         67         68         69         70 
#> 0.10691327 0.10040844 0.10527571 0.11763223 0.11485054 0.15945263 0.09935625 
#>         71         72         73         74         75         76         77 
#> 0.13289787 0.14345830 0.17530606 0.10876469 0.14479083 0.12994318 0.10489854 
#>         78         79         80         81         82         83         84 
#> 0.12609329 0.13344083 0.10720150 0.09902080 0.14188424 0.14952539 0.11678358 
#>         85         86         87         88         89         90         91 
#> 0.13728856 0.13215628 0.09981003 0.11608728 0.15888301 0.10709409 0.10225104 
#>         92         93         94         95         96         97         98 
#> 0.11975583 0.17279871 0.13035522 0.11998977 0.11717534 0.11162871 0.12045592 
#>         99        100        101        102        103        104        105 
#> 0.10543396 0.13201678 0.13093726 0.10471564 0.11823355 0.13061833 0.14706399 
#>        106        107        108        109        110        111        112 
#> 0.10225104 0.14597958 0.10901592 0.10607496 0.10478636 0.14188424 0.13008920 
#>        113        114        115        116        117        118        119 
#> 0.11084698 0.10760393 0.10607496 0.09895469 0.09835822 0.13616500 0.14613871 
#>        120        121        122        123        124        125        126 
#> 0.13061833 0.12658499 0.17353121 0.11823355 0.15161517 0.16195651 0.12111222 
#>        127        128        129        130        131        132        133 
#> 0.12766424 0.09767200 0.11115427 0.09866285 0.09831069 0.10097320 0.13008920 
#>        134        135        136        137        138        139        140 
#> 0.10478636 0.12763779 0.11010848 0.13611068 0.12391053 0.12873314 0.11695335 
#>        141        142        143        144        145        146        147 
#> 0.11608728 0.12830769 0.13749280 0.11678358 0.10097320 0.10522254 0.09806929 
#>        148        149        150        151        152        153        154 
#> 0.10213213 0.10537345 0.11572295 0.10499386 0.10853036 0.11240675 0.12790798 
#>        155        156        157        158        159        160        161 
#> 0.12163370 0.14461482 0.10471564 0.11132724 0.15945263 0.14594676 0.14594676 
#>        162        163        164        165        166        167        168 
#> 0.14602738 0.16148120 0.09767200 0.14955799 0.14430005 0.14955799 0.14600886 
#>        169        170        171        172        173        174        175 
#> 0.15319754 0.11975583 0.15697762 0.14602738 0.14714764 0.10158770 0.15319754 
#>        176        177        178        179        180        181        182 
#> 0.14714764 0.14723951 0.17837925 0.14447131 0.14936885 0.10522254 0.11291100 
#>        183        184        185        186        187        188        189 
#> 0.16700825 0.15888301 0.14723951 0.15626557 0.11483903 0.14823555 0.12357097 
#>        190        191        192        193        194        195        196 
#> 0.11231358 0.14897937 0.15063631 0.17186502 0.11247957 0.16786163 0.13155188 
#>        197        198        199        200        201        202        203 
#> 0.14811849 0.10522254 0.11586967 0.15697762 0.14600886 0.17061657 0.15945263 
#>        204        205        206        207        208        209        210 
#> 0.14662872 0.09806929 0.15530046 0.13932685 0.16405119 0.11214485 0.14510128 
#>        211        212        213        214        215        216        217 
#> 0.14427757 0.15710871 0.12287952 0.14662872 0.14936885 0.14811849 0.18478774 
#>        218        219        220        221        222        223        224 
#> 0.10901592 0.12111222 0.14467881 0.09767200 0.11516970 0.16939063 0.11480659 
#>        225        226        227        228        229        230        231 
#> 0.10087131 0.13580122 0.15052930 0.16175714 0.14781017 0.16786163 0.11115427 
#>        232        233        234        235        236        237        238 
#> 0.14467881 0.14638354 0.15319754 0.12695156 0.13292324 0.17186502 0.16584369 
#>        239        240        241        242        243        244        245 
#> 0.10663220 0.17186502 0.16257428 0.12681656 0.14440396 0.15173089 0.10527571 
#>        246        247        248        249        250        251        252 
#> 0.11615373 0.14479083 0.09943149 0.10489854 0.14510128 0.15104476 0.15052930 
#>        253        254        255        256        257        258        259 
#> 0.10097320 0.14946187 0.12982307 0.14525763 0.14913353 0.14658671 0.14781017 
#>        260        261        262        263        264        265        266 
#> 0.14709182 0.09791082 0.14714764 0.12743978 0.09935625 0.17447902 0.15661788 
#>        267        268        269        270        271        272        273 
#> 0.16175714 0.15185262 0.10522254 0.14625493 0.15069763 0.14600886 0.16987373 
#>        274        275        276        277        278        279        280 
#> 0.15063838 0.14746107 0.11480659 0.15319754 0.14622685 0.15063838 0.15888301 
#>        281        282        283        284        285        286        287 
#> 0.14994766 0.14936885 0.14897937 0.15063631 0.13161665 0.14973745 0.14811849 
#>        288        289        290        291        292        293        294 
#> 0.09835822 0.14440396 0.14823555 0.15436936 0.10518861 0.15847946 0.11932741 
#>        295        296        297        298        299        300        301 
#> 0.14625493 0.16786163 0.14367513 0.15573048 0.14600886 0.15697762 0.14906892 
#>        302        303        304        305        306        307        308 
#> 0.11480659 0.14897937 0.16472188 0.14662872 0.12274139 0.14662872 0.14430005 
#>        309        310        311        312        313        314 
#> 0.14802720 0.10760393 0.15063631 0.10703570 0.15984595 0.14936885 
#> 

# }