Initial guesses for mixed Poisson regression models

Function providing initial values for the optimization algorithms for mixed Poisson regression models.

startvalues_mpreg(y, x, w, link.mean, link.precision, model)

Arguments

y
response vector with z_i >= 0 and integers.

x
matrix containing the covariates for the mean submodel. Each column is a different covariate.

w
matrix containing the covariates for the precision submodel. Each column is a different covariate.

link.mean
a string containing the link function for the mean. The possible link functions for the mean are "log" and "sqrt".

link.precision
a string containing the link function the precision parameter. The possible link functions for the precision parameter are "identity", "log" or "inverse.sqrt".

model
the model, "NB" or "PIG". Currently not used.

Value

a list containing

beta - the initial guesses for the mean-related parameters;

alpha - the initial guesses for the precision-relared parameters.

@noRd