Fitted Method for mixpoissonreg Objects

Function providing the fitted means, linear predictors, precisions or variances for mixed Poisson regression models.

# S3 method for mixpoissonreg
fitted(object, type = c("response", "link", "precision", "variance"), ...)

Arguments

object
object of class "mixpoissonreg" containing results from the fitted model.

type
the type of variable to get the fitted values. The default is the "response" type, which provided the estimated values for the means. The type "link" provides the estimates for the linear predictor of the mean. The type "precision" provides estimates for the precision parameters whereas the type "variance" provides estimates for the variances.

...
Currently not used.

See also

Examples

# \donttest{
data("Attendance", package = "mixpoissonreg")

daysabs_fit <- mixpoissonreg(daysabs ~ gender + math +
prog | gender + math + prog, data = Attendance)
fitted(daysabs_fit)

#>         1         2         3         4         5         6         7         8 
#>  5.249599  6.661627  8.914762  9.153869 10.042385  6.361356  6.707173  7.808178 
#>         9        10        11        12        13        14        15        16 
#>  5.683434  2.477300  3.565489  6.885713  6.277724  9.361031  3.206667  4.419867 
#>        17        18        19        20        21        22        23        24 
#>  8.126015  8.072422  7.326732  7.756682  8.855968  5.319534  6.072228  5.249599 
#>        25        26        27        28        29        30        31        32 
#>  8.973947  8.474855  7.654707  8.855968  5.011988  7.230408  7.912199  8.399361 
#>        33        34        35        36        37        38        39        40 
#>  6.112542  6.153122  5.078758  8.914762  7.454760  4.568544 12.114724  8.072422 
#>        41        42        43        44        45        46        47        48 
#>  8.234269  8.126015  7.809716  6.574047  6.977444  8.511257  4.881071  6.193973 
#>        49        50        51        52        53        54        55        56 
#>  6.931427  5.954049  7.164589  5.390400  5.797383  7.071790  9.910359  5.462211 
#>        57        58        59        60        61        62        63        64 
#> 11.798278 11.955454  8.739540  8.072422  6.402328  8.511257  7.308234  5.874616 
#>        65        66        67        68        69        70        71        72 
#>  8.343966  6.444833 13.118487  9.275816  3.565489  7.407052  4.946096  4.722223 
#>        73        74        75        76        77        78        79        80 
#>  3.102311  6.795188  3.041933 14.296911  6.360104  5.112475  7.912199  8.855968 
#>        81        82        83        84        85        86        87        88 
#>  8.179963  4.753573  8.937326  5.390400  4.848879 10.543026  7.358202  9.337398 
#>        89        90        91        92        93        94        95        96 
#>  1.965161  6.661627  8.511257  6.403589  4.219812  9.804823  9.524607  7.260035 
#>        97        98        99       100       101       102       103       104 
#>  6.977444  7.405594  6.032181  7.860016 10.404418  6.235094  7.308234  7.808178 
#>       105       106       107       108       109       110       111       112 
#>  7.674081  8.511257 15.478431  6.796526  5.952877  6.318158  4.753573  9.976154 
#>       113       114       115       116       117       118       119       120 
#>  6.931427  6.705853  5.952877  7.456227  8.072422 14.777836  5.722293  7.808178 
#>       121       122       123       124       125       126       127       128 
#>  7.654707  6.767458  7.308234  4.568544  4.390717  6.361356 14.108952  7.758210 
#>       129       130       131       132       133       134       135       136 
#>  6.707173  8.126015  7.555559  8.399361  9.976154  6.318158  9.867974  6.885713 
#>       137       138       139       140       141       142       143       144 
#>  5.954049  6.277724  5.045262 13.031969  9.337398  6.154334 14.875945  5.390400 
#>       145       146       147       148       149       150       151       152 
#>  8.399361  6.978818  7.605720  7.166000  8.739540 12.860640  6.112542  5.759149 
#>       153       154       155       156       157       158       159       160 
#> 12.117110  9.933487 13.559772 11.566379  6.235094  5.608715  3.565489  2.349572 
#>       161       162       163       164       165       166       167       168 
#>  2.349572  2.318683  3.612988  7.758210  2.141690  2.962475  2.141690  2.754506 
#>       169       170       171       172       173       174       175       176 
#>  2.058326  6.403589  2.461446  2.318683  2.228430  8.455124  2.058326  2.228430 
#>       177       178       179       180       181       182       183       184 
#>  2.700365  4.136871  3.001941  2.629830  6.978818 11.416566  3.734523  1.965161 
#>       185       186       187       188       189       190       191       192 
#>  2.700365  2.004561 11.045029  2.664864  5.180583 11.645461  2.560632  2.595256 
#>       193       194       195       196       197       198       199       200 
#>  3.834688  9.153869  5.285491  9.740159  2.526968  6.978818  6.531976  2.461446 
#>       201       202       203       204       205       206       207       208 
#>  2.754506  3.809398  3.565489  2.258117  7.605720  2.735801  9.361031  2.350035 
#>       209       210       211       212       213       214       215       216 
#> 11.878944  2.809732  2.923528  1.991341 10.269655  2.258117  2.629830  2.526968 
#>       217       218       219       220       221       222       223       224 
#>  3.249386  6.796526  6.361356  2.847163  7.758210  7.164589  3.784274  5.462211 
#>       225       226       227       228       229       230       231       232 
#>  7.261465  4.881071  2.611971  1.926535  2.199133  5.285491  6.707173  2.847163 
#>       233       234       235       236       237       238       239       240 
#>  2.736339  2.058326 10.001404  6.033369  3.834688  6.995104  6.617692  3.834688 
#>       241       242       243       244       245       246       247       248 
#>  2.884525  6.195192  2.885093  2.646768  6.444833  7.212154  3.041933  8.234269 
#>       249       250       251       252       253       254       255       256 
#>  6.360104  2.809732  3.337197  2.611971  8.399361  2.577632  6.113745  3.082457 
#>       257       258       259       260       261       262       263       264 
#>  3.271603  2.444731  2.199133  2.477300  7.966296  2.228430  9.175231  7.407052 
#>       265       266       267       268       269       270       271       272 
#>  8.092854  3.495408  1.926535  2.085747  6.978818  2.288199  7.523245  2.754506 
#>       273       274       275       276       277       278       279       280 
#>  2.273556  2.113533  2.213733  5.462211  2.058326  2.412591  2.113533  1.965161 
#>       281       282       283       284       285       286       287       288 
#>  4.598875  2.629830  2.560632  2.595256  8.587757  3.293323  2.526968  8.072422 
#>       289       290       291       292       293       294       295       296 
#>  2.885093  2.664864  2.510797  6.072228  3.541974  7.356753  2.288199  5.285491 
#>       297       298       299       300       301       302       303       304 
#>  7.827941  3.472355  2.754506  2.461446  2.155908  5.462211  2.560632  3.685426 
#>       305       306       307       308       309       310       311       312 
#>  2.258117  9.651493  2.258117  2.962475  3.228592  6.705853  2.595256  6.887069 
#>       313       314 
#>  2.413066  2.629830 

fitted(daysabs_fit, type = "precision")

#>         1         2         3         4         5         6         7         8 
#> 0.8177972 0.9831495 1.2553572 1.2813073 1.3764305 0.9670989 1.0074950 1.1115667 
#>         9        10        11        12        13        14        15        16 
#> 0.8695681 0.6267233 0.8465645 1.0086188 0.9572556 2.7685650 0.7650931 0.7159566 
#>        17        18        19        20        21        22        23        24 
#> 1.1686012 1.1626389 2.2472815 1.1058954 1.2489522 0.8262066 0.9152055 0.8177972 
#>        25        26        27        28        29        30        31        32 
#> 1.2617950 2.5149563 1.0946393 1.2489522 0.7890330 2.2244082 1.1229969 1.1988748 
#>        33        34        35        36        37        38        39        40 
#> 0.9198989 0.9246164 0.7971465 1.2553572 1.0724698 0.7345040 3.3150597 1.1626389 
#>        41        42        43        44        45        46        47        48 
#> 1.1806178 1.1686012 1.1332803 0.9731428 1.0189903 1.2112027 0.7730528 0.9293581 
#>        49        50        51        52        53        54        55        56 
#> 1.0137913 0.9188739 1.0400544 0.8347023 0.8830151 1.0495784 1.3624209 0.8432855 
#>        57        58        59        60        61        62        63        64 
#> 3.2479202 3.2813182 1.2362401 1.1626389 0.9534339 1.2112027 1.0561377 0.8920950 
#>        65        66        67        68        69        70        71        72 
#> 1.1927580 0.9583234 3.5937771 1.2944829 0.8465645 1.0878408 0.7810020 0.7535319 
#>        73        74        75        76        77        78        79        80 
#> 0.7457732 0.9983529 0.7487625 3.8408789 0.9485694 0.8012345 1.1229969 1.2489522 
#>        81        82        83        84        85        86        87        88 
#> 1.1745942 0.7573963 2.6711866 0.8347023 0.7691086 2.9774185 1.0822905 1.3011213 
#>        89        90        91        92        93        94        95        96 
#> 0.5239905 0.9831495 1.2112027 0.9720585 0.6907743 2.8694932 1.3212418 1.0507492 
#>        97        98        99       100       101       102       103       104 
#> 1.0189903 1.0669979 0.9105360 1.1172672 2.9471137 0.9341241 1.0561377 1.1115667 
#>       105       106       107       108       109       110       111       112 
#> 2.3292063 1.2112027 4.0840269 1.0178549 0.9012683 0.9437296 0.7573963 1.3694078 
#>       113       114       115       116       117       118       119       120 
#> 1.0137913 0.9881914 0.9012683 1.0934196 1.1626389 3.9403800 0.8911010 1.1115667 
#>       121       122       123       124       125       126       127       128 
#> 1.0946393 2.1134866 1.0561377 0.7345040 0.7123037 0.9670989 3.8017856 1.1274982 
#>       129       130       131       132       133       134       135       136 
#> 1.0074950 1.1686012 1.1046631 1.1988748 1.3694078 0.9437296 2.8289476 1.0086188 
#>       137       138       139       140       141       142       143       144 
#> 0.9188739 0.9572556 0.7930794 3.5754412 1.3011213 0.9426781 3.9605874 0.8347023 
#>       145       146       147       148       149       150       151       152 
#> 1.1988748 1.0388955 1.1103282 1.0603711 1.2362401 3.5390495 0.9198989 0.8785098 
#>       153       154       155       156       157       158       159       160 
#> 3.3798168 2.8434553 3.6868768 3.1984595 0.9341241 0.8607175 0.8465645 0.6015945 
#>       161       162       163       164       165       166       167       168 
#> 0.6015945 0.5954713 0.8552696 1.1274982 0.5600191 0.7335979 0.5600191 0.6934603 
#>       169       170       171       172       173       174       175       176 
#> 0.5430926 0.9720585 0.6357058 0.5954713 0.5774732 1.2050230 0.5430926 0.5774732 
#>       177       178       179       180       181       182       183       184 
#> 0.6829000 0.6802549 0.7411414 0.6690693 1.0388955 3.2277484 0.8774261 0.5239905 
#>       185       186       187       188       189       190       191       192 
#> 0.6829000 0.5320934 3.1462424 0.6759493 0.8094735 3.2776620 0.6429591 0.6622594 
#>       193       194       195       196       197       198       199       200 
#> 0.8955638 1.2813073 0.8380481 2.8548527 0.6364149 1.0388955 0.9870903 0.6357058 
#>       201       202       203       204       205       206       207       208 
#> 0.6934603 0.8909945 0.8465645 0.5834113 1.1103282 0.6767033 2.7685650 0.6133462 
#>       209       210       211       212       213       214       215       216 
#> 3.3283475 0.7041839 0.7261312 0.5293786 2.9741009 0.5834113 0.6690693 0.6364149 
#>       217       218       219       220       221       222       223       224 
#> 0.7729604 1.0178549 0.9670989 0.7114250 1.1274982 1.0400544 0.8864485 0.8432855 
#>       225       226       227       228       229       230       231       232 
#> 1.0712747 0.7730528 0.6529017 0.5160109 0.5715956 0.8380481 1.0074950 0.7114250 
#>       233       234       235       236       237       238       239       240 
#> 0.6899222 0.5430926 2.9138668 0.9283226 0.8955638 2.1682382 0.9781334 0.8955638 
#>       241       242       243       244       245       246       247       248 
#> 0.7049694 0.9475124 0.7187405 0.6596154 0.9583234 1.0453881 0.7487625 1.1806178 
#>       249       250       251       252       253       254       255       256 
#> 0.9485694 0.7041839 0.8043500 0.6529017 1.1988748 0.6462564 0.9378684 0.7564619 
#>       257       258       259       260       261       262       263       264 
#> 0.7921010 0.6203444 0.5715956 0.6267233 1.1508053 0.5774732 2.6741663 1.0878408 
#>       265       266       267       268       269       270       271       272 
#> 2.4738978 0.8336726 0.5160109 0.5486772 1.0388955 0.5894105 2.2937362 0.6934603 
#>       273       274       275       276       277       278       279       280 
#> 0.5978582 0.5543191 0.5745269 0.8432855 0.5430926 0.6140304 0.5543191 0.5239905 
#>       281       282       283       284       285       286       287       288 
#> 0.7382708 0.6690693 0.6429591 0.6622594 2.5408173 0.7961631 0.6364149 1.1626389 
#>       289       290       291       292       293       294       295       296 
#> 0.7187405 0.6759493 0.6455362 0.9152055 0.8422452 1.0615539 0.5894105 0.8380481 
#>       297       298       299       300       301       302       303       304 
#> 2.3652250 0.8294191 0.6934603 0.6357058 0.5628911 0.8432855 0.6429591 0.8684954 
#>       305       306       307       308       309       310       311       312 
#> 0.5834113 1.3348280 0.5834113 0.7335979 0.7840388 0.9881914 0.6622594 1.0283214 
#>       313       314 
#> 0.6260250 0.6690693 

# }