Predict Method for mixpoissonreg Objects

Function to obtain various predictions based on the fitted mixed Poisson regression models.

# S3 method for mixpoissonreg
predict(
  object,
  newdata = NULL,
  type = c("response", "link", "precision", "variance"),
  se.fit = FALSE,
  interval = c("none", "confidence", "prediction"),
  level = 0.95,
  nsim_pred = 100,
  nsim_pred_y = 100,
  ...
)

Arguments

object
object of class "mixpoissonreg" containing results from the fitted model.

newdata
optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted response values will be provided.

type
the type of prediction. The default is the "response" type, which provided the estimated values for the means. The type "link" provides the estimates for the linear predictor. The type "precision" provides estimates for the precision parameters whereas the type "variance" provides estimates for the variances.

se.fit
logical switch indicating if standard errors on the scale of linear predictors should be returned. If TRUE, it only returns the standard deviations of the linear predictors when type = 'link', otherwise returns NA and a warning indicating that the type must be 'link'. When using mixpoissonreg objects, the fit must be done with x = TRUE.

interval
Type of interval calculation for the response variables, 'none', 'confidence' or 'prediction'. If 'confidence', the confidence intervals for the means are returned. If 'prediction', prediction intervals for future response variables are reported. For confidence intervals, the type of the prediction must be 'response' or 'link'. For prediction intervals the type of prediction must be 'response'. For 'confidence' intervals, when using mixpoissonreg objects, the fit must be done with x = TRUE and for predictions intervals, the fit must be done with x = TRUE and w = TRUE.

level
Tolerance/confidence level. The default is set to 0.95.

nsim_pred
number of means and predictions to be generated in each step of the simulation. The default is set to 100.

nsim_pred_y
number of response variables generated for each pair of mean and precision to compute the prediction intervals. The default is set to 100.

...
further arguments passed to or from other methods.

Details

The se.fit argument only returns a non-NA vector for type = 'link', that is, on the scale of the linear predictor for the mean parameter. For the response scale, one can obtain confidence or prediction intervals. It is important to notice that confidence intervals must not be used for future observations as they will underestimate the uncertainty. In this case prediction intervals should be used. Currently, we do not have closed-form expressions for the prediction interval and, therefore, they are obtained by simulation and can be computationally-intensive.

See also

Examples

# \donttest{
data("Attendance", package = "mixpoissonreg")

daysabs_fit <- mixpoissonreg(daysabs ~ gender + math +
prog | gender + math + prog, data = Attendance)
predict(daysabs_fit, interval = "confidence")

#>           fit       lwr       upr
#> 1    5.249554  4.139862  6.656699
#> 2    6.661619  5.400442  8.217322
#> 3    8.915540  7.212229 11.021121
#> 4    9.154674  7.343539 11.412490
#> 5   10.043298  7.759662 12.998998
#> 6    6.361844  5.017556  8.066289
#> 7    6.707698  5.394571  8.340463
#> 8    7.808208  6.044846 10.085968
#> 9    5.683400  4.582659  7.048535
#> 10   2.476910  1.856557  3.304547
#> 11   3.565250  2.608360  4.873180
#> 12   6.885712  5.549246  8.544051
#> 13   6.278203  4.924513  8.004007
#> 14   9.361429  7.125088 12.299689
#> 15   3.206188  2.244399  4.580131
#> 16   4.419806  3.228922  6.049907
#> 17   8.126700  6.697776  9.860473
#> 18   8.073101  6.657572  9.789599
#> 19   7.326407  5.398103  9.943538
#> 20   7.756711  6.021142  9.992549
#> 21   8.856738  7.178386 10.927500
#> 22   5.319490  4.213676  6.715508
#> 23   6.072204  4.941004  7.462383
#> 24   5.249554  4.139862  6.656699
#> 25   8.974731  7.245656 11.116426
#> 26   8.474518  6.531633 10.995329
#> 27   7.654731  5.973045  9.809889
#> 28   8.856738  7.178386 10.927500
#> 29   5.011938  3.883977  6.467475
#> 30   7.230085  5.298626  9.865601
#> 31   7.912232  6.091596 10.277014
#> 32   8.400077  6.891833 10.238395
#> 33   6.112518  4.975708  7.509059
#> 34   6.153100  5.010156  7.556779
#> 35   5.078709  3.956568  6.519106
#> 36   8.915540  7.212229 11.021121
#> 37   7.454777  5.873904  9.461119
#> 38   4.568486  3.394132  6.149159
#> 39  12.114378  8.975794 16.350436
#> 40   8.073101  6.657572  9.789599
#> 41   8.234967  6.776799 10.006889
#> 42   8.126700  6.697776  9.860473
#> 43   7.810365  6.449843  9.457873
#> 44   6.574037  5.338584  8.095398
#> 45   6.977447  5.606453  8.683701
#> 46   8.511986  6.966191 10.400794
#> 47   4.881018  3.740461  6.369359
#> 48   6.193952  5.044338  7.605564
#> 49   6.931428  5.578012  8.613228
#> 50   5.954494  4.560270  7.774976
#> 51   7.164597  5.717033  8.978688
#> 52   5.390359  4.287630  6.776697
#> 53   5.797352  4.691941  7.163195
#> 54   7.072355  5.772832  8.664415
#> 55   9.911256  7.703727 12.751361
#> 56   5.462171  4.361616  6.840426
#> 57  11.797931  8.827134 15.768557
#> 58  11.955107  8.902039 16.055264
#> 59   8.740297  7.109424 10.745286
#> 60   8.073101  6.657572  9.789599
#> 61   6.402313  5.210926  7.866090
#> 62   8.511986  6.966191 10.400794
#> 63   7.308247  5.796759  9.213852
#> 64   5.874586  4.764066  7.243973
#> 65   8.344676  6.853955 10.159625
#> 66   6.444819  5.243325  7.921631
#> 67  13.119184 10.419092 16.518999
#> 68   9.276636  7.406888 11.618373
#> 69   3.565250  2.608360  4.873180
#> 70   7.407655  6.096842  9.000292
#> 71   4.946045  3.811918  6.417598
#> 72   4.722167  3.564836  6.255227
#> 73   3.101844  2.199943  4.373493
#> 74   6.795184  5.490723  8.409554
#> 75   3.041714  2.290183  4.039862
#> 76  14.297708 11.084370 18.442587
#> 77   6.360087  5.178212  7.811714
#> 78   5.112427  3.993040  6.545618
#> 79   7.912232  6.091596 10.277014
#> 80   8.856738  7.178386 10.927500
#> 81   8.180654  6.737520  9.932898
#> 82   4.753518  3.599590  6.277364
#> 83   8.937693  6.667941 11.980064
#> 84   5.390359  4.287630  6.776697
#> 85   4.848826  3.704977  6.345819
#> 86  10.542678  8.137934 13.658021
#> 87   7.358799  6.051247  8.948888
#> 88   9.338225  7.438020 11.723880
#> 89   1.964837  1.439096  2.682647
#> 90   6.661619  5.400442  8.217322
#> 91   8.511986  6.966191 10.400794
#> 92   6.404081  5.064310  8.098290
#> 93   4.219747  3.007566  5.920490
#> 94   9.805255  7.595295 12.658234
#> 95   9.525457  7.529365 12.050727
#> 96   7.260047  5.770479  9.134126
#> 97   6.977447  5.606453  8.683701
#> 98   7.405611  5.848466  9.377343
#> 99   6.032156  4.906055  7.416733
#> 100  7.860048  6.068328 10.180787
#> 101 10.404071  8.050150 13.446296
#> 102  6.235074  5.078244  7.655433
#> 103  7.308247  5.796759  9.213852
#> 104  7.808208  6.044846 10.085968
#> 105  7.673752  5.752735 10.236256
#> 106  8.511986  6.966191 10.400794
#> 107 15.479332 11.629127 20.604275
#> 108  6.797061  5.489392  8.416239
#> 109  5.952850  4.835474  7.328427
#> 110  6.318140  5.145189  7.758491
#> 111  4.753518  3.599590  6.277364
#> 112  9.977059  7.731820 12.874291
#> 113  6.931428  5.578012  8.613228
#> 114  6.705847  5.430870  8.280143
#> 115  5.952850  4.835474  7.328427
#> 116  7.456836  6.142152  9.052918
#> 117  8.073101  6.657572  9.789599
#> 118 14.778675 11.318380 19.296865
#> 119  5.722712  4.297481  7.620613
#> 120  7.808208  6.044846 10.085968
#> 121  7.654731  5.973045  9.809889
#> 122  6.767141  4.816608  9.507562
#> 123  7.308247  5.796759  9.213852
#> 124  4.568486  3.394132  6.149159
#> 125  4.390655  3.196577  6.030781
#> 126  6.361844  5.017556  8.066289
#> 127 14.109733 10.987608 18.119009
#> 128  7.758852  6.407017  9.395916
#> 129  6.707698  5.394571  8.340463
#> 130  8.126700  6.697776  9.860473
#> 131  7.556179  6.231843  9.161951
#> 132  8.400077  6.891833 10.238395
#> 133  9.977059  7.731820 12.874291
#> 134  6.318140  5.145189  7.758491
#> 135  9.867629  7.684599 12.670811
#> 136  6.885712  5.549246  8.544051
#> 137  5.954494  4.560270  7.774976
#> 138  6.278203  4.924513  8.004007
#> 139  5.045213  3.920210  6.493063
#> 140 13.032658 10.364154 16.388233
#> 141  9.338225  7.438020 11.723880
#> 142  6.154799  4.786300  7.914581
#> 143 14.876793 11.363913 19.475596
#> 144  5.390359  4.287630  6.776697
#> 145  8.400077  6.891833 10.238395
#> 146  6.979373  5.678685  8.577981
#> 147  7.606345  6.276187  9.218414
#> 148  7.166576  5.866420  8.754880
#> 149  8.740297  7.109424 10.745286
#> 150 12.861315 10.252585 16.133825
#> 151  6.112518  4.975708  7.509059
#> 152  5.759116  4.655645  7.124130
#> 153 12.117723  9.722787 15.102584
#> 154  9.933141  7.731524 12.761687
#> 155 13.560506 10.685459 17.209117
#> 156 11.566032  8.712517 15.354128
#> 157  6.235074  5.078244  7.655433
#> 158  5.608679  4.509247  6.976171
#> 159  3.565250  2.608360  4.873180
#> 160  2.349198  1.764785  3.127142
#> 161  2.349198  1.764785  3.127142
#> 162  2.318313  1.741307  3.086517
#> 163  3.612748  2.632626  4.957767
#> 164  7.758852  6.407017  9.395916
#> 165  2.141343  1.597293  2.870699
#> 166  2.962260  2.232504  3.930556
#> 167  2.141343  1.597293  2.870699
#> 168  2.754299  2.068834  3.666879
#> 169  2.057990  1.524209  2.778702
#> 170  6.404081  5.064310  8.098290
#> 171  2.461253  1.809409  3.347925
#> 172  2.318313  1.741307  3.086517
#> 173  2.228071  1.669856  2.972893
#> 174  8.455847  6.929244 10.318781
#> 175  2.057990  1.524209  2.778702
#> 176  2.228071  1.669856  2.972893
#> 177  2.700161  2.023284  3.603483
#> 178  4.136805  2.916383  5.867937
#> 179  3.001724  2.261488  3.984256
#> 180  2.629629  1.962229  3.524027
#> 181  6.979373  5.678685  8.577981
#> 182 11.417123  9.151571 14.243531
#> 183  3.734278  2.691878  5.180337
#> 184  1.964837  1.439096  2.682647
#> 185  2.700161  2.023284  3.603483
#> 186  2.004232  1.475497  2.722436
#> 187 11.045556  8.820322 13.832183
#> 188  2.664662  1.992791  3.563054
#> 189  5.180537  4.066286  6.600117
#> 190 11.646036  9.346009 14.512092
#> 191  2.560232  1.911928  3.428364
#> 192  2.595056  1.931628  3.486343
#> 193  3.834440  2.737903  5.370143
#> 194  9.154674  7.343539 11.412490
#> 195  5.285866  3.804014  7.344973
#> 196  9.740586  7.527522 12.604281
#> 197  2.526572  1.889976  3.377589
#> 198  6.979373  5.678685  8.577981
#> 199  6.532483  5.205337  8.197996
#> 200  2.461253  1.809409  3.347925
#> 201  2.754299  2.068834  3.666879
#> 202  3.809151  2.726506  5.321692
#> 203  3.565250  2.608360  4.873180
#> 204  2.257755  1.693824  3.009437
#> 205  7.606345  6.276187  9.218414
#> 206  2.735378  2.017584  3.708542
#> 207  9.361429  7.125088 12.299689
#> 208  2.349847  1.703730  3.240994
#> 209 11.879538  9.536561 14.798147
#> 210  2.809523  2.114071  3.733753
#> 211  2.923314  2.203249  3.878710
#> 212  1.991014  1.463346  2.708952
#> 213 10.270122  8.072834 13.065475
#> 214  2.257755  1.693824  3.009437
#> 215  2.629629  1.962229  3.524027
#> 216  2.526572  1.889976  3.377589
#> 217  3.248902  2.261830  4.666738
#> 218  6.797061  5.489392  8.416239
#> 219  6.361844  5.017556  8.066289
#> 220  2.846952  2.144010  3.780364
#> 221  7.758852  6.407017  9.395916
#> 222  7.164597  5.717033  8.978688
#> 223  3.784028  2.715038  5.273911
#> 224  5.462171  4.361616  6.840426
#> 225  7.262051  5.959281  8.849623
#> 226  4.881018  3.740461  6.369359
#> 227  2.611564  1.944350  3.507737
#> 228  1.926217  1.402886  2.644770
#> 229  2.198778  1.645763  2.937619
#> 230  5.285866  3.804014  7.344973
#> 231  6.707698  5.394571  8.340463
#> 232  2.846952  2.144010  3.780364
#> 233  2.736134  2.053681  3.645370
#> 234  2.057990  1.524209  2.778702
#> 235 10.001850  7.799447 12.826165
#> 236  6.033822  4.649976  7.829504
#> 237  3.834440  2.737903  5.370143
#> 238  6.994784  5.054221  9.680425
#> 239  6.617683  5.369679  8.155744
#> 240  3.834440  2.737903  5.370143
#> 241  2.884084  2.097167  3.966276
#> 242  6.195662  4.832175  7.943882
#> 243  2.884881  2.173744  3.828665
#> 244  2.646356  1.965620  3.562846
#> 245  6.444819  5.243325  7.921631
#> 246  7.212164  5.743905  9.055740
#> 247  3.041714  2.290183  4.039862
#> 248  8.234967  6.776799 10.006889
#> 249  6.360087  5.178212  7.811714
#> 250  2.809523  2.114071  3.733753
#> 251  3.336966  2.481893  4.486633
#> 252  2.611564  1.944350  3.507737
#> 253  8.400077  6.891833 10.238395
#> 254  2.577229  1.922804  3.454389
#> 255  6.114206  4.740635  7.885762
#> 256  3.082237  2.318573  4.097427
#> 257  3.271375  2.442236  4.382007
#> 258  2.444345  1.833962  3.257877
#> 259  2.198778  1.645763  2.937619
#> 260  2.476910  1.856557  3.304547
#> 261  7.966962  6.575800  9.652436
#> 262  2.228071  1.669856  2.972893
#> 263  9.174889  7.147862 11.776749
#> 264  7.407655  6.096842  9.000292
#> 265  8.093161  5.749678 11.391813
#> 266  3.495172  2.571331  4.750935
#> 267  1.926217  1.402886  2.644770
#> 268  2.085408  1.548591  2.808311
#> 269  6.979373  5.678685  8.577981
#> 270  2.287833  1.717647  3.047297
#> 271  7.522918  5.599633 10.106785
#> 272  2.754299  2.068834  3.666879
#> 273  2.273372  1.629585  3.171494
#> 274  2.113190  1.572960  2.838961
#> 275  2.213376  1.657824  2.955100
#> 276  5.462171  4.361616  6.840426
#> 277  2.057990  1.524209  2.778702
#> 278  2.412209  1.811127  3.212779
#> 279  2.113190  1.572960  2.838961
#> 280  1.964837  1.439096  2.682647
#> 281  4.598816  3.427849  6.169791
#> 282  2.629629  1.962229  3.524027
#> 283  2.560232  1.911928  3.428364
#> 284  2.595056  1.931628  3.486343
#> 285  8.587419  6.635611 11.113335
#> 286  3.293094  2.455544  4.416321
#> 287  2.526572  1.889976  3.377589
#> 288  8.073101  6.657572  9.789599
#> 289  2.884881  2.173744  3.828665
#> 290  2.664662  1.992791  3.563054
#> 291  2.510601  1.855144  3.397643
#> 292  6.072204  4.941004  7.462383
#> 293  3.541736  2.596102  4.831819
#> 294  7.356768  5.822753  9.294922
#> 295  2.287833  1.717647  3.047297
#> 296  5.285866  3.804014  7.344973
#> 297  7.827611  5.907189 10.372359
#> 298  3.472120  2.558817  4.711403
#> 299  2.754299  2.068834  3.666879
#> 300  2.461253  1.809409  3.347925
#> 301  2.155560  1.609440  2.886990
#> 302  5.462171  4.361616  6.840426
#> 303  2.560232  1.911928  3.428364
#> 304  3.685183  2.668414  5.089381
#> 305  2.257755  1.693824  3.009437
#> 306  9.652358  7.588654 12.277279
#> 307  2.257755  1.693824  3.009437
#> 308  2.962260  2.232504  3.930556
#> 309  3.228366  2.415357  4.315033
#> 310  6.705847  5.430870  8.280143
#> 311  2.595056  1.931628  3.486343
#> 312  6.887614  5.584150  8.495334
#> 313  2.412875  1.763902  3.300617
#> 314  2.629629  1.962229  3.524027

predict(daysabs_fit, type = "link", se.fit = TRUE)

#> $fit
#>         1         2         3         4         5         6         7         8 
#> 1.6581431 1.8963626 2.1877958 2.2142646 2.3069055 1.8503182 1.9032559 2.0551755 
#>         9        10        11        12        13        14        15        16 
#> 1.7375496 0.9070117 1.2712342 1.9294486 1.8370838 2.2365980 1.1650828 1.4860957 
#>        17        18        19        20        21        22        23        24 
#> 2.0951549 2.0885377 1.9914852 2.0485583 2.1811786 1.6713775 1.8037217 1.6581431 
#>        25        26        27        28        29        30        31        32 
#> 2.1944130 2.1370638 2.0353239 2.1811786 1.6118227 1.9782508 2.0684100 2.1282409 
#>        33        34        35        36        37        38        39        40 
#> 1.8103389 1.8169561 1.6250571 2.1877958 2.0088551 1.5191818 2.4943930 2.0885377 
#>        41        42        43        44        45        46        47        48 
#> 2.1083893 2.0951549 2.0554516 1.8831282 1.9426830 2.1414753 1.5853538 1.8235733 
#>        49        50        51        52        53        54        55        56 
#> 1.9360658 1.7841461 1.9691519 1.6846119 1.7574012 1.9561935 2.2936711 1.6978464 
#>        57        58        59        60        61        62        63        64 
#> 2.4679242 2.4811586 2.1679442 2.0885377 1.8566593 2.1414753 1.9890035 1.7706356 
#>        65        66        67        68        69        70        71        72 
#> 2.1216237 1.8632765 2.5740756 2.2274990 1.2712342 2.0025140 1.5985882 1.5522678 
#>        73        74        75        76        77        78        79        80 
#> 1.1319967 1.9162142 1.1124212 2.6600993 1.8500421 1.6316743 2.0684100 2.1811786 
#>        81        82        83        84        85        86        87        88 
#> 2.1017721 1.5588850 2.1902775 1.6846119 1.5787366 2.3554316 1.9958968 2.2341162 
#>        89        90        91        92        93        94        95        96 
#> 0.6754094 1.8963626 2.1414753 1.8569354 1.4397753 2.2829184 2.2539679 1.9823863 
#>        97        98        99       100       101       102       103       104 
#> 1.9426830 2.0022379 1.7971045 2.0617928 2.3421972 1.8301905 1.9890035 2.0551755 
#>       105       106       107       108       109       110       111       112 
#> 2.0378057 2.1414753 2.7395057 1.9164903 1.7838701 1.8434249 1.5588850 2.3002883 
#>       113       114       115       116       117       118       119       120 
#> 1.9360658 1.9029798 1.7838701 2.0091312 2.0885377 2.6931853 1.7444429 2.0551755 
#>       121       122       123       124       125       126       127       128 
#> 2.0353239 1.9120787 1.9890035 1.5191818 1.4794785 1.8503182 2.6468648 2.0488344 
#>       129       130       131       132       133       134       135       136 
#> 1.9032559 2.0951549 2.0223656 2.1282409 2.3002883 1.8434249 2.2892596 1.9294486 
#>       137       138       139       140       141       142       143       144 
#> 1.7841461 1.8370838 1.6184399 2.5674584 2.2341162 1.8172322 2.6998025 1.6846119 
#>       145       146       147       148       149       150       151       152 
#> 2.1282409 1.9429591 2.0289828 1.9694279 2.1679442 2.5542239 1.8103389 1.7507840 
#>       153       154       155       156       157       158       159       160 
#> 2.4946691 2.2958768 2.6071616 2.4480725 1.8301905 1.7243152 1.2712342 0.8540740 
#>       161       162       163       164       165       166       167       168 
#> 0.8540740 0.8408396 1.2844686 2.0488344 0.7614331 1.0859524 0.7614331 1.0131631 
#>       169       170       171       172       173       174       175       176 
#> 0.7217299 1.8569354 0.9006706 0.8408396 0.8011364 2.1348581 0.7217299 0.8011364 
#>       177       178       179       180       181       182       183       184 
#> 0.9933115 1.4199237 1.0991868 0.9668426 1.9429591 2.4351142 1.3175546 0.6754094 
#>       185       186       187       188       189       190       191       192 
#> 0.9933115 0.6952610 2.4020282 0.9800770 1.6449087 2.4549658 0.9400977 0.9536082 
#>       193       194       195       196       197       198       199       200 
#> 1.3440234 2.2142646 1.6650364 2.2763012 0.9268633 1.9429591 1.8767870 0.9006706 
#>       201       202       203       204       205       206       207       208 
#> 1.0131631 1.3374062 1.2712342 0.8143708 2.0289828 1.0062698 2.2365980 0.8543501 
#>       209       210       211       212       213       214       215       216 
#> 2.4748175 1.0330147 1.0727179 0.6886438 2.3292389 0.8143708 0.9668426 0.9268633 
#>       217       218       219       220       221       222       223       224 
#> 1.1783172 1.9164903 1.8503182 1.0462491 2.0488344 1.9691519 1.3307890 1.6978464 
#>       225       226       227       228       229       230       231       232 
#> 1.9826624 1.5853538 0.9599493 0.6555578 0.7879019 1.6650364 1.9032559 1.0462491 
#>       233       234       235       236       237       238       239       240 
#> 1.0065459 0.7217299 2.3027701 1.7973806 1.3440234 1.9451648 1.8897454 1.3440234 
#>       241       242       243       244       245       246       247       248 
#> 1.0592074 1.8238494 1.0594835 0.9731837 1.8632765 1.9757691 1.1124212 2.1083893 
#>       249       250       251       252       253       254       255       256 
#> 1.8500421 1.0330147 1.2050621 0.9599493 2.1282409 0.9467149 1.8106150 1.1256556 
#>       257       258       259       260       261       262       263       264 
#> 1.1852105 0.8937773 0.7879019 0.9070117 2.0753033 0.8011364 2.2164703 2.0025140 
#>       265       266       267       268       269       270       271       272 
#> 2.0910194 1.2513825 0.6555578 0.7349643 1.9429591 0.8276052 2.0179541 1.0131631 
#>       273       274       275       276       277       278       279       280 
#> 0.8212641 0.7481987 0.7945191 1.6978464 0.7217299 0.8805428 0.7481987 0.6754094 
#>       281       282       283       284       285       286       287       288 
#> 1.5257990 0.9668426 0.9400977 0.9536082 2.1502982 1.1918277 0.9268633 2.0885377 
#>       289       290       291       292       293       294       295       296 
#> 1.0594835 0.9800770 0.9205222 1.8037217 1.2646170 1.9956207 0.8276052 1.6650364 
#>       297       298       299       300       301       302       303       304 
#> 2.0576573 1.2447653 1.0131631 0.9006706 0.7680503 1.6978464 0.9400977 1.3043202 
#>       305       306       307       308       309       310       311       312 
#> 0.8143708 2.2672023 0.8143708 1.0859524 1.1719761 1.9029798 0.9536082 1.9297247 
#>       313       314 
#> 0.8808189 0.9668426 
#> 
#> $se.fit
#>          1          2          3          4          5          6          7 
#> 0.12116582 0.10708454 0.10817432 0.11247337 0.13161810 0.12111206 0.11115656 
#>          8          9         10         11         12         13         14 
#> 0.13059912 0.10983377 0.14708830 0.15944813 0.11009723 0.12390966 0.13927602 
#>         15         16         17         18         19         20         21 
#> 0.18196498 0.16018011 0.09866471 0.09836039 0.15583842 0.12922756 0.10719788 
#>         22         23         24         25         26         27         28 
#> 0.11890116 0.10518212 0.12116582 0.10919121 0.13286305 0.12656713 0.10719788 
#>         29         30         31         32         33         34         35 
#> 0.13008558 0.15857597 0.13342071 0.10097323 0.10498724 0.10484324 0.12739015 
#>         36         37         38         39         40         41         42 
#> 0.10817432 0.12160203 0.15160154 0.15299341 0.09836039 0.09943268 0.09866471 
#>         43         44         45         46         47         48         49 
#> 0.09765272 0.10620992 0.11161672 0.10225024 0.13579078 0.10475030 0.11083536 
#>         50         51         52         53         54         55         56 
#> 0.13610671 0.11515613 0.11677659 0.10793817 0.10358907 0.12855692 0.11479988 
#>         57         58         59         60         61         62         63 
#> 0.14800973 0.15045084 0.10537081 0.09836039 0.10505383 0.10225024 0.11821874 
#>         64         65         66         67         68         69         70 
#> 0.10690713 0.10040888 0.10526757 0.11757137 0.11484327 0.15944813 0.09936054 
#>         71         72         73         74         75         76         77 
#> 0.13288800 0.14344633 0.17529164 0.10875415 0.14479319 0.12988159 0.10489087 
#>         78         79         80         81         82         83         84 
#> 0.12608471 0.13342071 0.10719788 0.09902233 0.14187258 0.14947540 0.11677659 
#>         85         86         87         88         89         90         91 
#> 0.13727783 0.13209188 0.09981432 0.11607947 0.15887771 0.10708454 0.10225024 
#>         92         93         94         95         96         97         98 
#> 0.11975601 0.17278073 0.13030315 0.11998032 0.11716094 0.11161672 0.12044025 
#>         99        100        101        102        103        104        105 
#> 0.10542757 0.13199711 0.13087305 0.10470857 0.11821874 0.13059912 0.14700794 
#>        106        107        108        109        110        111        112 
#> 0.10225024 0.14591739 0.10901871 0.10606874 0.10477891 0.14187258 0.13007587 
#>        113        114        115        116        117        118        119 
#> 0.11083536 0.10759406 0.10606874 0.09895896 0.09836039 0.13610319 0.14613223 
#>        120        121        122        123        124        125        126 
#> 0.13059912 0.12656713 0.17347704 0.11821874 0.15160154 0.16194076 0.12111206 
#>        127        128        129        130        131        132        133 
#> 0.12760273 0.09767562 0.11115656 0.09866471 0.09831483 0.10097323 0.13007587 
#>        134        135        136        137        138        139        140 
#> 0.10477891 0.12757450 0.11009723 0.13610671 0.12390966 0.12872407 0.11689257 
#>        141        142        143        144        145        146        147 
#> 0.11607947 0.12830570 0.13743094 0.11677659 0.10097323 0.10522612 0.09807333 
#>        148        149        150        151        152        153        154 
#> 0.10213623 0.10537081 0.11566236 0.10498724 0.10852425 0.11234734 0.12784456 
#>        155        156        157        158        159        160        161 
#> 0.12157249 0.14454949 0.10470857 0.11132095 0.15944813 0.14594417 0.14594417 
#>        162        163        164        165        166        167        168 
#> 0.14602490 0.16147588 0.09767562 0.14955506 0.14430287 0.14955506 0.14601177 
#>        169        170        171        172        173        174        175 
#> 0.15319373 0.11975601 0.15697749 0.14602490 0.14714517 0.10158732 0.15319373 
#>        176        177        178        179        180        181        182 
#> 0.14714517 0.14724215 0.17836013 0.14447393 0.14937093 0.10522612 0.11285340 
#>        183        184        185        186        187        188        189 
#> 0.16700081 0.15887771 0.14724215 0.15626098 0.11478263 0.14823793 0.12356285 
#>        190        191        192        193        194        195        196 
#> 0.11225532 0.14897488 0.15063803 0.17185579 0.11247337 0.16784986 0.13150001 
#>        197        198        199        200        201        202        203 
#> 0.14811442 0.10522612 0.11587082 0.15697749 0.14601177 0.17060780 0.15944813 
#>        204        205        206        207        208        209        210 
#> 0.14662629 0.09807333 0.15529324 0.13927602 0.16404894 0.11208598 0.14510434 
#>        211        212        213        214        215        216        217 
#> 0.14428054 0.15710389 0.12282589 0.14662629 0.14937093 0.14811442 0.18477029 
#>        218        219        220        221        222        223        224 
#> 0.10901871 0.12111206 0.14468189 0.09767562 0.11515613 0.16938230 0.11479988 
#>        225        226        227        228        229        230        231 
#> 0.10087555 0.13579078 0.15052409 0.16175106 0.14780759 0.16784986 0.11115656 
#>        232        233        234        235        236        237        238 
#> 0.14468189 0.14638637 0.15319373 0.12689885 0.13292008 0.17185579 0.16578926 
#>        239        240        241        242        243        244        245 
#> 0.10662295 0.17185579 0.16256428 0.12681495 0.14440701 0.15172515 0.10526757 
#>        246        247        248        249        250        251        252 
#> 0.11613975 0.14479319 0.09943268 0.10489087 0.14510434 0.15104384 0.15052409 
#>        253        254        255        256        257        258        259 
#> 0.10097323 0.14945714 0.12982069 0.14525967 0.14913351 0.14658350 0.14780759 
#>        260        261        262        263        264        265        266 
#> 0.14708830 0.09791357 0.14714517 0.12737832 0.09936054 0.17442950 0.15661453 
#>        267        268        269        270        271        272        273 
#> 0.16175106 0.15184915 0.10522612 0.14625250 0.15064203 0.14601177 0.16986973 
#>        274        275        276        277        278        279        280 
#> 0.15063520 0.14745855 0.11479988 0.15319373 0.14622390 0.15063520 0.15887771 
#>        281        282        283        284        285        286        287 
#> 0.14993437 0.14937093 0.14897488 0.15063803 0.13155726 0.14973714 0.14811442 
#>        288        289        290        291        292        293        294 
#> 0.09836039 0.14440701 0.14823793 0.15437001 0.10518212 0.15847535 0.11931217 
#>        295        296        297        298        299        300        301 
#> 0.14625250 0.16784986 0.14361856 0.15572751 0.14601177 0.15697749 0.14906610 
#>        302        303        304        305        306        307        308 
#> 0.11479988 0.14897488 0.16471530 0.14662629 0.12273083 0.14662629 0.14430287 
#>        309        310        311        312        313        314 
#> 0.14802772 0.10759406 0.15063803 0.10703892 0.15984496 0.14937093 
#> 

# }