Predict Method for mixpoissonreg Objects

Function to obtain various predictions based on the fitted mixed Poisson regression models.

# S3 method for mixpoissonreg
predict(
  object,
  newdata = NULL,
  type = c("response", "link", "precision", "variance"),
  se.fit = FALSE,
  interval = c("none", "confidence", "prediction"),
  level = 0.95,
  nsim_pred = 100,
  nsim_pred_y = 100,
  ...
)

Arguments

object
object of class "mixpoissonreg" containing results from the fitted model.

newdata
optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted response values will be provided.

type
the type of prediction. The default is the "response" type, which provided the estimated values for the means. The type "link" provides the estimates for the linear predictor. The type "precision" provides estimates for the precision parameters whereas the type "variance" provides estimates for the variances.

se.fit
logical switch indicating if standard errors on the scale of linear predictors should be returned. If TRUE, it only returns the standard deviations of the linear predictors when type = 'link', otherwise returns NA and a warning indicating that the type must be 'link'. When using mixpoissonreg objects, the fit must be done with x = TRUE.

interval
Type of interval calculation for the response variables, 'none', 'confidence' or 'prediction'. If 'confidence', the confidence intervals for the means are returned. If 'prediction', prediction intervals for future response variables are reported. For confidence intervals, the type of the prediction must be 'response' or 'link'. For prediction intervals the type of prediction must be 'response'. For 'confidence' intervals, when using mixpoissonreg objects, the fit must be done with x = TRUE and for predictions intervals, the fit must be done with x = TRUE and w = TRUE.

level
Tolerance/confidence level. The default is set to 0.95.

nsim_pred
number of means and predictions to be generated in each step of the simulation. The default is set to 100.

nsim_pred_y
number of response variables generated for each pair of mean and precision to compute the prediction intervals. The default is set to 100.

...
further arguments passed to or from other methods.

Details

The se.fit argument only returns a non-NA vector for type = 'link', that is, on the scale of the linear predictor for the mean parameter. For the response scale, one can obtain confidence or prediction intervals. It is important to notice that confidence intervals must not be used for future observations as they will underestimate the uncertainty. In this case prediction intervals should be used. Currently, we do not have closed-form expressions for the prediction interval and, therefore, they are obtained by simulation and can be computationally-intensive.

See also

Examples

# \donttest{
data("Attendance", package = "mixpoissonreg")

daysabs_fit <- mixpoissonreg(daysabs ~ gender + math +
prog | gender + math + prog, data = Attendance)
predict(daysabs_fit, interval = "confidence")

#>           fit       lwr       upr
#> 1    5.249599  4.139831  6.656863
#> 2    6.661627  5.400335  8.217503
#> 3    8.914762  7.211531 11.020266
#> 4    9.153869  7.342789 11.411647
#> 5   10.042385  7.758715 12.998221
#> 6    6.361356  5.017171  8.065671
#> 7    6.707173  5.394177  8.339766
#> 8    7.808178  6.044565 10.086359
#> 9    5.683434  4.582628  7.048669
#> 10   2.477300  1.856836  3.305092
#> 11   3.565489  2.608507  4.873558
#> 12   6.885713  5.549109  8.544263
#> 13   6.277724  4.924129  8.003409
#> 14   9.361031  7.123984 12.300547
#> 15   3.206667  2.244652  4.580984
#> 16   4.419867  3.228862  6.050190
#> 17   8.126015  6.697238  9.859604
#> 18   8.072422  6.657042  9.788732
#> 19   7.326732  5.397689  9.945181
#> 20   7.756682  6.020870  9.992928
#> 21   8.855968  7.177701 10.926642
#> 22   5.319534  4.213646  6.715665
#> 23   6.072228  4.940955  7.462516
#> 24   5.249599  4.139831  6.656863
#> 25   8.973947  7.244946 11.115573
#> 26   8.474855  6.531045 10.997193
#> 27   7.654707  5.972789  9.810246
#> 28   8.855968  7.177701 10.926642
#> 29   5.011988  3.883940  6.467666
#> 30   7.230408  5.298225  9.867230
#> 31   7.912199  6.091298 10.277429
#> 32   8.399361  6.891243 10.237524
#> 33   6.112542  4.975657  7.509193
#> 34   6.153122  5.010102  7.556915
#> 35   5.078758  3.956533  6.519288
#> 36   8.914762  7.211531 11.020266
#> 37   7.454760  5.873680  9.461434
#> 38   4.568544  3.394078  6.149416
#> 39  12.114724  8.974746 16.353281
#> 40   8.072422  6.657042  9.788732
#> 41   8.234269  6.776240 10.006018
#> 42   8.126015  6.697238  9.859604
#> 43   7.809716  6.449355  9.457017
#> 44   6.574047  5.338488  8.095569
#> 45   6.977444  5.606303  8.683927
#> 46   8.511257  6.965579 10.399924
#> 47   4.881071  3.740419  6.369568
#> 48   6.193973  5.044281  7.605702
#> 49   6.931427  5.577869  8.613447
#> 50   5.954049  4.559892  7.774461
#> 51   7.164589  5.716855  8.978946
#> 52   5.390400  4.287600  6.776848
#> 53   5.797383  4.691906  7.163326
#> 54   7.071790  5.772420  8.663647
#> 55   9.910359  7.702809 12.750573
#> 56   5.462211  4.361587  6.840571
#> 57  11.798278  8.826116 15.771305
#> 58  11.955454  8.901005 16.058060
#> 59   8.739540  7.108764 10.744423
#> 60   8.072422  6.657042  9.788732
#> 61   6.402328  5.210849  7.866242
#> 62   8.511257  6.965579 10.399924
#> 63   7.308234  5.796559  9.214137
#> 64   5.874616  4.764029  7.244103
#> 65   8.343966  6.853376 10.158754
#> 66   6.444833  5.243244  7.921788
#> 67  13.118487 10.417133 16.520352
#> 68   9.275816  7.406111 11.617537
#> 69   3.565489  2.608507  4.873558
#> 70   7.407052  6.096403  8.999475
#> 71   4.946096  3.811879  6.417798
#> 72   4.722223  3.564788  6.255459
#> 73   3.102311  2.200204  4.374291
#> 74   6.795188  5.490599  8.409752
#> 75   3.041933  2.290359  4.040134
#> 76  14.296911 11.082238 18.444077
#> 77   6.360104  5.178139  7.811864
#> 78   5.112475  3.993006  6.545796
#> 79   7.912199  6.091298 10.277429
#> 80   8.855968  7.177701 10.926642
#> 81   8.179963  6.736971  9.932028
#> 82   4.753573  3.599543  6.277592
#> 83   8.937326  6.666931 11.980894
#> 84   5.390400  4.287600  6.776848
#> 85   4.848879  3.704934  6.346033
#> 86  10.543026  8.137044 13.660413
#> 87   7.358202  6.050812  8.948077
#> 88   9.337398  7.437229 11.723049
#> 89   1.965161  1.439318  2.683117
#> 90   6.661627  5.400335  8.217503
#> 91   8.511257  6.965579 10.399924
#> 92   6.403589  5.063925  8.097662
#> 93   4.219812  3.007497  5.920809
#> 94   9.804823  7.594086 12.659134
#> 95   9.524607  7.528533 12.049908
#> 96   7.260035  5.770286  9.134402
#> 97   6.977444  5.606303  8.683927
#> 98   7.405594  5.848250  9.377648
#> 99   6.032181  4.906010  7.416865
#> 100  7.860016  6.068039 10.181190
#> 101 10.404418  8.049277 13.448651
#> 102  6.235094  5.078183  7.655573
#> 103  7.308234  5.796559  9.214137
#> 104  7.808178  6.044565 10.086359
#> 105  7.674081  5.752273 10.237955
#> 106  8.511257  6.965579 10.399924
#> 107 15.478431 11.626847 20.605915
#> 108  6.796526  5.488995  8.415523
#> 109  5.952877  4.835433  7.328558
#> 110  6.318158  5.145120  7.758637
#> 111  4.753573  3.599543  6.277592
#> 112  9.976154  7.730888 12.873508
#> 113  6.931427  5.577869  8.613447
#> 114  6.705853  5.430757  8.280330
#> 115  5.952877  4.835433  7.328558
#> 116  7.456227  6.141708  9.052095
#> 117  8.072422  6.657042  9.788732
#> 118 14.777836 11.316186 19.298413
#> 119  5.722293  4.297106  7.620159
#> 120  7.808178  6.044565 10.086359
#> 121  7.654707  5.972789  9.810246
#> 122  6.767458  4.816261  9.509136
#> 123  7.308234  5.796559  9.214137
#> 124  4.568544  3.394078  6.149416
#> 125  4.390717  3.196515  6.031069
#> 126  6.361356  5.017171  8.065671
#> 127 14.108952 10.985501 18.120477
#> 128  7.758210  6.406536  9.395064
#> 129  6.707173  5.394177  8.339766
#> 130  8.126015  6.697238  9.859604
#> 131  7.555559  6.231388  9.161117
#> 132  8.399361  6.891243 10.237524
#> 133  9.976154  7.730888 12.873508
#> 134  6.318158  5.145120  7.758637
#> 135  9.867974  7.683794 12.673024
#> 136  6.885713  5.549109  8.544263
#> 137  5.954049  4.559892  7.774461
#> 138  6.277724  4.924129  8.003409
#> 139  5.045262  3.920174  6.493250
#> 140 13.031969 10.362209 16.389575
#> 141  9.337398  7.437229 11.723049
#> 142  6.154334  4.785919  7.914014
#> 143 14.875945 11.361707 19.477157
#> 144  5.390400  4.287600  6.776848
#> 145  8.399361  6.891243 10.237524
#> 146  6.978818  5.678279  8.577230
#> 147  7.605720  6.275726  9.217575
#> 148  7.166000  5.866002  8.754097
#> 149  8.739540  7.108764 10.744423
#> 150 12.860640 10.250670 16.135146
#> 151  6.112542  4.975657  7.509193
#> 152  5.759149  4.655611  7.124261
#> 153 12.117110  9.721015 15.103808
#> 154  9.933487  7.730711 12.763917
#> 155 13.559772 10.683431 17.210521
#> 156 11.566379  8.711520 15.356807
#> 157  6.235094  5.078183  7.655573
#> 158  5.608715  4.509217  6.976308
#> 159  3.565489  2.608507  4.873558
#> 160  2.349572  1.765057  3.127655
#> 161  2.349572  1.765057  3.127655
#> 162  2.318683  1.741577  3.087024
#> 163  3.612988  2.632768  4.958159
#> 164  7.758210  6.406536  9.395064
#> 165  2.141690  1.597543  2.871180
#> 166  2.962475  2.232680  3.930818
#> 167  2.141690  1.597543  2.871180
#> 168  2.754506  2.069002  3.667131
#> 169  2.058326  1.524447  2.779177
#> 170  6.403589  5.063925  8.097662
#> 171  2.461446  1.809551  3.348189
#> 172  2.318683  1.741577  3.087024
#> 173  2.228430  1.670116  2.973385
#> 174  8.455124  6.928643 10.317911
#> 175  2.058326  1.524447  2.779177
#> 176  2.228430  1.670116  2.973385
#> 177  2.700365  2.023449  3.603735
#> 178  4.136871  2.916310  5.868271
#> 179  3.001941  2.261664  3.984522
#> 180  2.629830  1.962388  3.524280
#> 181  6.978818  5.678279  8.577230
#> 182 11.416566  9.149958 14.244654
#> 183  3.734523  2.692008  5.180766
#> 184  1.965161  1.439318  2.683117
#> 185  2.700365  2.023449  3.603735
#> 186  2.004561  1.475726  2.722907
#> 187 11.045029  8.818799 13.833251
#> 188  2.664864  1.992953  3.563307
#> 189  5.180583  4.066253  6.600288
#> 190 11.645461  9.344342 14.513249
#> 191  2.560632  1.912209  3.428933
#> 192  2.595256  1.931784  3.486598
#> 193  3.834688  2.738022  5.370604
#> 194  9.153869  7.342789 11.411647
#> 195  5.285491  3.803647  7.344639
#> 196  9.740159  7.526330 12.605174
#> 197  2.526968  1.890257  3.378148
#> 198  6.978818  5.678279  8.577230
#> 199  6.531976  5.204948  8.197338
#> 200  2.461446  1.809551  3.348189
#> 201  2.754506  2.069002  3.667131
#> 202  3.809398  2.726628  5.322145
#> 203  3.565489  2.608507  4.873558
#> 204  2.258117  1.694088  3.009933
#> 205  7.605720  6.275726  9.217575
#> 206  2.735801  2.017863  3.709174
#> 207  9.361031  7.123984 12.300547
#> 208  2.350035  1.703858  3.241269
#> 209 11.878944  9.534840 14.799337
#> 210  2.809732  2.114242  3.734007
#> 211  2.923528  2.203424  3.878970
#> 212  1.991341  1.463573  2.709423
#> 213 10.269655  8.071509 13.066431
#> 214  2.258117  1.694088  3.009933
#> 215  2.629830  1.962388  3.524280
#> 216  2.526968  1.890257  3.378148
#> 217  3.249386  2.262079  4.667614
#> 218  6.796526  5.488995  8.415523
#> 219  6.361356  5.017171  8.065671
#> 220  2.847163  2.144182  3.780619
#> 221  7.758210  6.406536  9.395064
#> 222  7.164589  5.716855  8.978946
#> 223  3.784274  2.715162  5.274355
#> 224  5.462211  4.361587  6.840571
#> 225  7.261465  5.958856  8.848825
#> 226  4.881071  3.740419  6.369568
#> 227  2.611971  1.944631  3.508323
#> 228  1.926535  1.403101  2.645239
#> 229  2.199133  1.646021  2.938107
#> 230  5.285491  3.803647  7.344639
#> 231  6.707173  5.394177  8.339766
#> 232  2.847163  2.144182  3.780619
#> 233  2.736339  2.053848  3.645622
#> 234  2.058326  1.524447  2.779177
#> 235 10.001404  7.798190 12.827088
#> 236  6.033369  4.649596  7.828968
#> 237  3.834688  2.738022  5.370604
#> 238  6.995104  5.053849  9.682023
#> 239  6.617692  5.369578  8.155920
#> 240  3.834688  2.738022  5.370604
#> 241  2.884525  2.097441  3.966969
#> 242  6.195192  4.831793  7.943305
#> 243  2.885093  2.173918  3.828922
#> 244  2.646768  1.965902  3.563444
#> 245  6.444833  5.243244  7.921788
#> 246  7.212154  5.743720  9.056007
#> 247  3.041933  2.290359  4.040134
#> 248  8.234269  6.776240 10.006018
#> 249  6.360104  5.178139  7.811864
#> 250  2.809732  2.114242  3.734007
#> 251  3.337197  2.482058  4.486955
#> 252  2.611971  1.944631  3.508323
#> 253  8.399361  6.891243 10.237524
#> 254  2.577632  1.923085  3.454963
#> 255  6.113745  4.740254  7.885206
#> 256  3.082457  2.318748  4.097704
#> 257  3.271603  2.442405  4.382315
#> 258  2.444731  1.834240  3.258412
#> 259  2.199133  1.646021  2.938107
#> 260  2.477300  1.856836  3.305092
#> 261  7.966296  6.575288  9.651572
#> 262  2.228430  1.670116  2.973385
#> 263  9.175231  7.147161 11.778784
#> 264  7.407052  6.096403  8.999475
#> 265  8.092854  5.748832 11.392624
#> 266  3.495408  2.571484  4.751294
#> 267  1.926535  1.403101  2.645239
#> 268  2.085747  1.548833  2.808787
#> 269  6.978818  5.678279  8.577230
#> 270  2.288199  1.717914  3.047798
#> 271  7.523245  5.599193 10.108459
#> 272  2.754506  2.069002  3.667131
#> 273  2.273556  1.629703  3.171777
#> 274  2.113533  1.573206  2.839440
#> 275  2.213733  1.658083  2.955590
#> 276  5.462211  4.361587  6.840571
#> 277  2.058326  1.524447  2.779177
#> 278  2.412591  1.811403  3.213307
#> 279  2.113533  1.573206  2.839440
#> 280  1.965161  1.439318  2.683117
#> 281  4.598875  3.427796  6.170043
#> 282  2.629830  1.962388  3.524280
#> 283  2.560632  1.912209  3.428933
#> 284  2.595256  1.931784  3.486598
#> 285  8.587757  6.635004 11.115225
#> 286  3.293323  2.455711  4.416634
#> 287  2.526968  1.890257  3.378148
#> 288  8.072422  6.657042  9.788732
#> 289  2.885093  2.173918  3.828922
#> 290  2.664864  1.992953  3.563307
#> 291  2.510797  1.855292  3.397902
#> 292  6.072228  4.940955  7.462516
#> 293  3.541974  2.596251  4.832190
#> 294  7.356753  5.822545  9.295217
#> 295  2.288199  1.717914  3.047798
#> 296  5.285491  3.803647  7.344639
#> 297  7.827941  5.906704 10.374087
#> 298  3.472355  2.558972  4.711756
#> 299  2.754506  2.069002  3.667131
#> 300  2.461446  1.809551  3.348189
#> 301  2.155908  1.609692  2.887473
#> 302  5.462211  4.361587  6.840571
#> 303  2.560632  1.912209  3.428933
#> 304  3.685426  2.668549  5.089794
#> 305  2.258117  1.694088  3.009933
#> 306  9.651493  7.587793 12.276470
#> 307  2.258117  1.694088  3.009933
#> 308  2.962475  2.232680  3.930818
#> 309  3.228592  2.415528  4.315333
#> 310  6.705853  5.430757  8.280330
#> 311  2.595256  1.931784  3.486598
#> 312  6.887069  5.583749  8.494601
#> 313  2.413066  1.764038  3.300884
#> 314  2.629830  1.962388  3.524280

predict(daysabs_fit, type = "link", se.fit = TRUE)

#> $fit
#>         1         2         3         4         5         6         7         8 
#> 1.6581517 1.8963637 2.1877086 2.2141766 2.3068146 1.8502416 1.9031776 2.0551717 
#>         9        10        11        12        13        14        15        16 
#> 1.7375557 0.9071692 1.2713011 1.9294487 1.8370076 2.2365554 1.1652322 1.4861096 
#>        17        18        19        20        21        22        23        24 
#> 2.0950706 2.0884536 1.9915295 2.0485547 2.1810916 1.6713857 1.8037257 1.6581517 
#>        25        26        27        28        29        30        31        32 
#> 2.1943256 2.1371035 2.0353207 2.1810916 1.6118326 1.9782955 2.0684057 2.1281556 
#>        33        34        35        36        37        38        39        40 
#> 1.8103427 1.8169597 1.6250666 2.1877086 2.0088527 1.5191946 2.4944216 2.0884536 
#>        41        42        43        44        45        46        47        48 
#> 2.1083046 2.0950706 2.0553686 1.8831297 1.9426827 2.1413896 1.5853646 1.8235767 
#>        49        50        51        52        53        54        55        56 
#> 1.9360657 1.7840716 1.9691507 1.6846197 1.7574067 1.9561136 2.2935806 1.6978537 
#>        57        58        59        60        61        62        63        64 
#> 2.4679536 2.4811876 2.1678576 2.0884536 1.8566617 2.1413896 1.9890017 1.7706407 
#>        65        66        67        68        69        70        71        72 
#> 2.1215386 1.8632787 2.5740225 2.2274106 1.2713011 2.0024326 1.5985986 1.5522796 
#>        73        74        75        76        77        78        79        80 
#> 1.1321472 1.9162147 1.1124931 2.6600435 1.8500447 1.6316836 2.0684057 2.1810916 
#>        81        82        83        84        85        86        87        88 
#> 2.1016876 1.5588966 2.1902364 1.6846197 1.5787476 2.3554646 1.9958156 2.2340276 
#>        89        90        91        92        93        94        95        96 
#> 0.6755741 1.8963637 2.1413896 1.8568586 1.4397906 2.2828744 2.2538786 1.9823847 
#>        97        98        99       100       101       102       103       104 
#> 1.9426827 2.0022357 1.7971087 2.0617887 2.3422306 1.8301937 1.9890017 2.0551717 
#>       105       106       107       108       109       110       111       112 
#> 2.0378485 2.1413896 2.7394475 1.9164116 1.7838747 1.8434277 1.5588966 2.3001976 
#>       113       114       115       116       117       118       119       120 
#> 1.9360657 1.9029807 1.7838747 2.0090496 2.0884536 2.6931285 1.7443696 2.0551717 
#>       121       122       123       124       125       126       127       128 
#> 2.0353207 1.9121255 1.9890017 1.5191946 1.4794926 1.8502416 2.6468095 2.0487516 
#>       129       130       131       132       133       134       135       136 
#> 1.9031776 2.0950706 2.0222836 2.1281556 2.3001976 1.8434277 2.2892945 1.9294487 
#>       137       138       139       140       141       142       143       144 
#> 1.7840716 1.8370076 1.6184496 2.5674055 2.2340276 1.8171566 2.6997455 1.6846197 
#>       145       146       147       148       149       150       151       152 
#> 2.1281556 1.9428796 2.0289006 1.9693476 2.1678576 2.5541715 1.8103427 1.7507897 
#>       153       154       155       156       157       158       159       160 
#> 2.4946185 2.2959115 2.6071075 2.4481026 1.8301937 1.7243217 1.2713011 0.8542332 
#>       161       162       163       164       165       166       167       168 
#> 0.8542332 0.8409992 1.2845351 2.0487516 0.7615951 1.0860251 0.7615951 1.0132381 
#>       169       170       171       172       173       174       175       176 
#> 0.7218931 1.8568586 0.9007491 0.8409992 0.8012972 2.1347726 0.7218931 0.8012972 
#>       177       178       179       180       181       182       183       184 
#> 0.9933871 1.4199396 1.0992591 0.9669191 1.9428796 2.4350655 1.3176201 0.6755741 
#>       185       186       187       188       189       190       191       192 
#> 0.9933871 0.6954251 2.4019805 0.9801531 1.6449177 2.4549165 0.9402542 0.9536851 
#>       193       194       195       196       197       198       199       200 
#> 1.3440881 2.2141766 1.6649655 2.2762574 0.9270202 1.9428796 1.8767096 0.9007491 
#>       201       202       203       204       205       206       207       208 
#> 1.0132381 1.3374711 1.2713011 0.8145312 2.0289006 1.0064242 2.2365554 0.8544301 
#>       209       210       211       212       213       214       215       216 
#> 2.4747675 1.0330891 1.0727911 0.6888081 2.3291934 0.8145312 0.9669191 0.9270202 
#>       217       218       219       220       221       222       223       224 
#> 1.1784662 1.9164116 1.8502416 1.0463231 2.0487516 1.9691507 1.3308541 1.6978537 
#>       225       226       227       228       229       230       231       232 
#> 1.9825816 1.5853646 0.9601052 0.6557231 0.7880632 1.6649655 1.9031776 1.0463231 
#>       233       234       235       236       237       238       239       240 
#> 1.0066211 0.7218931 2.3027254 1.7973056 1.3440881 1.9452105 1.8897467 1.3440881 
#>       241       242       243       244       245       246       247       248 
#> 1.0593602 1.8237736 1.0595571 0.9733392 1.8632787 1.9757677 1.1124931 2.1083046 
#>       249       250       251       252       253       254       255       256 
#> 1.8500447 1.0330891 1.2051311 0.9601052 2.1281556 0.9468712 1.8105396 1.1257271 
#>       257       258       259       260       261       262       263       264 
#> 1.1852801 0.8939352 0.7880632 0.9071692 2.0752196 0.8012972 2.2165075 2.0024326 
#>       265       266       267       268       269       270       271       272 
#> 2.0909814 1.2514501 0.6557231 0.7351271 1.9428796 0.8277652 2.0179975 1.0132381 
#>       273       274       275       276       277       278       279       280 
#> 0.8213450 0.7483611 0.7946802 1.6978537 0.7218931 0.8807012 0.7483611 0.6755741 
#>       281       282       283       284       285       286       287       288 
#> 1.5258116 0.9669191 0.9402542 0.9536851 2.1503375 1.1918971 0.9270202 2.0884536 
#>       289       290       291       292       293       294       295       296 
#> 1.0595571 0.9801531 0.9206001 1.8037257 1.2646841 1.9956187 0.8277652 1.6649655 
#>       297       298       299       300       301       302       303       304 
#> 2.0576995 1.2448331 1.0132381 0.9007491 0.7682121 1.6978537 0.9402542 1.3043861 
#>       305       306       307       308       309       310       311       312 
#> 0.8145312 2.2671126 0.8145312 1.0860251 1.1720461 1.9029807 0.9536851 1.9296456 
#>       313       314 
#> 0.8808981 0.9669191 
#> 
#> $se.fit
#>          1          2          3          4          5          6          7 
#> 0.12117401 0.10709520 0.10817918 0.11248058 0.13163399 0.12111209 0.11115389 
#>          8          9         10         11         12         13         14 
#> 0.13062087 0.10984034 0.14709202 0.15945355 0.11010985 0.12391047 0.13933332 
#>         15         16         17         18         19         20         21 
#> 0.18198374 0.16019682 0.09866275 0.09835808 0.15590012 0.12924879 0.10720218 
#>         22         23         24         25         26         27         28 
#> 0.11890894 0.10518916 0.12117401 0.10919665 0.13292926 0.12658734 0.10720218 
#>         29         30         31         32         33         34         35 
#> 0.13009556 0.15863741 0.13344349 0.10097337 0.10499444 0.10485059 0.12739957 
#>         36         37         38         39         40         41         42 
#> 0.10817918 0.12162023 0.15161629 0.15306759 0.09835808 0.09943150 0.09866275 
#>         43         44         45         46         47         48         49 
#> 0.09764896 0.10621988 0.11163019 0.10225133 0.13580199 0.10475784 0.11084840 
#>         50         51         52         53         54         55         56 
#> 0.13611096 0.11517142 0.11678401 0.10794469 0.10358467 0.12857155 0.11480700 
#>         57         58         59         60         61         62         63 
#> 0.14808362 0.15052487 0.10537399 0.09835808 0.10506256 0.10225133 0.11823546 
#>         64         65         66         67         68         69         70 
#> 0.10691371 0.10040856 0.10527657 0.11764022 0.11485170 0.15945355 0.09935574 
#>         71         72         73         74         75         76         77 
#> 0.13289857 0.14345924 0.17530794 0.10876595 0.14479075 0.12995127 0.10489932 
#>         78         79         80         81         82         83         84 
#> 0.12609386 0.13344349 0.10720218 0.09902075 0.14188514 0.14953170 0.11678401 
#>         85         86         87         88         89         90         91 
#> 0.13728936 0.13216445 0.09980951 0.11608851 0.15888307 0.10709520 0.10225133 
#>         92         93         94         95         96         97         98 
#> 0.11975566 0.17280033 0.13036189 0.11999121 0.11717718 0.11163019 0.12045795 
#>         99        100        101        102        103        104        105 
#> 0.10543448 0.13201938 0.13094540 0.10471631 0.11823546 0.13062087 0.14707077 
#>        106        107        108        109        110        111        112 
#> 0.10225133 0.14598773 0.10901549 0.10607545 0.10478710 0.14188514 0.13009114 
#>        113        114        115        116        117        118        119 
#> 0.11084840 0.10760509 0.10607545 0.09895420 0.09835808 0.13617311 0.14613929 
#>        120        121        122        123        124        125        126 
#> 0.13062087 0.12658734 0.17353765 0.11823546 0.15161629 0.16195788 0.12111209 
#>        127        128        129        130        131        132        133 
#> 0.12767231 0.09767164 0.11115389 0.09866275 0.09831022 0.10097337 0.13009114 
#>        134        135        136        137        138        139        140 
#> 0.10478710 0.12764575 0.11010985 0.13611096 0.12391047 0.12873376 0.11696133 
#>        141        142        143        144        145        146        147 
#> 0.11608851 0.12830775 0.13750091 0.11678401 0.10097337 0.10522204 0.09806884 
#>        148        149        150        151        152        153        154 
#> 0.10213160 0.10537399 0.11573091 0.10499444 0.10853077 0.11241453 0.12791597 
#>        155        156        157        158        159        160        161 
#> 0.12164174 0.14462319 0.10471631 0.11132764 0.15945355 0.14594675 0.14594675 
#>        162        163        164        165        166        167        168 
#> 0.14602733 0.16148223 0.09767164 0.14955787 0.14429988 0.14955787 0.14600859 
#>        169        170        171        172        173        174        175 
#> 0.15319747 0.11975566 0.15697758 0.14602733 0.14714753 0.10158792 0.15319747 
#>        176        177        178        179        180        181        182 
#> 0.14714753 0.14723925 0.17838101 0.14447118 0.14936862 0.10522204 0.11291851 
#>        183        184        185        186        187        188        189 
#> 0.16700956 0.15888307 0.14723925 0.15626557 0.11484635 0.14823530 0.12357150 
#>        190        191        192        193        194        195        196 
#> 0.11232119 0.14897974 0.15063611 0.17186657 0.11248058 0.16786282 0.13155852 
#>        197        198        199        200        201        202        203 
#> 0.14811878 0.10522204 0.11586940 0.15697758 0.14600859 0.17061806 0.15945355 
#>        204        205        206        207        208        209        210 
#> 0.14662862 0.09806884 0.15530128 0.13933332 0.16405137 0.11215255 0.14510102 
#>        211        212        213        214        215        216        217 
#> 0.14427737 0.15710873 0.12288643 0.14662862 0.14936862 0.14811878 0.18479005 
#>        218        219        220        221        222        223        224 
#> 0.10901549 0.12111209 0.14467856 0.09767164 0.11517142 0.16939206 0.11480700 
#>        225        226        227        228        229        230        231 
#> 0.10087078 0.13580199 0.15052979 0.16175727 0.14781005 0.16786282 0.11115389 
#>        232        233        234        235        236        237        238 
#> 0.14467856 0.14638327 0.15319747 0.12695832 0.13292344 0.17186657 0.16585019 
#>        239        240        241        242        243        244        245 
#> 0.10663325 0.17186657 0.16257553 0.12681658 0.14440373 0.15173147 0.10527657 
#>        246        247        248        249        250        251        252 
#> 0.11615551 0.14479075 0.09943150 0.10489932 0.14510102 0.15104518 0.15052979 
#>        253        254        255        256        257        258        259 
#> 0.10097337 0.14946227 0.12982317 0.14525761 0.14913383 0.14658684 0.14781005 
#>        260        261        262        263        264        265        266 
#> 0.14709202 0.09791060 0.14714753 0.12744743 0.09935574 0.17448520 0.15661864 
#>        267        268        269        270        271        272        273 
#> 0.16175727 0.15185253 0.10522204 0.14625485 0.15070434 0.14600859 0.16987409 
#>        274        275        276        277        278        279        280 
#> 0.15063826 0.14746095 0.11480700 0.15319747 0.14622692 0.15063826 0.15888307 
#>        281        282        283        284        285        286        287 
#> 0.14994874 0.14936862 0.14897974 0.15063611 0.13162396 0.14973779 0.14811878 
#>        288        289        290        291        292        293        294 
#> 0.09835808 0.14440373 0.14823530 0.15436926 0.10518916 0.15848033 0.11932938 
#>        295        296        297        298        299        300        301 
#> 0.14625485 0.16786282 0.14368199 0.15573119 0.14600859 0.15697758 0.14906879 
#>        302        303        304        305        306        307        308 
#> 0.11480700 0.14897974 0.16472308 0.14662862 0.12274297 0.14662862 0.14429988 
#>        309        310        311        312        313        314 
#> 0.14802741 0.10760509 0.15063611 0.10703524 0.15984599 0.14936862 
#> 

# }