Calculate Variance-Covariance Matrix for mixpoissonreg Objects

Returns the variance-covariance matrix of the parameters for fitted mixed Poisson regression models. The parameters argument indicates for which parameters the variance-covariance matrix should be computed, namely, 'mean' for mean-relatex parameters or 'precision' for precision-related parameters.

# S3 method for mixpoissonreg
vcov(object, parameters = c("all", "mean", "precision"), ...)

Arguments

object
an object of class "mixpoissonreg" containing results from the fitted model.

parameters
a string to determine which coefficients should be extracted: 'all' extracts all coefficients, 'mean' extracts the coefficients of the mean parameters and 'precision' extracts coefficients of the precision parameters.

...
further arguments passed to or from other methods.

See also

Examples

# \donttest{
data("Attendance", package = "mixpoissonreg")

daysabs_fit <- mixpoissonreg(daysabs ~ gender + math +
prog | gender + math + prog, data = Attendance)
vcov(daysabs_fit)

#>                            (Intercept)    gendermale          math
#> (Intercept)               2.173024e-02 -5.918840e-03 -2.218622e-04
#> gendermale               -5.918840e-03  1.391630e-02  1.824148e-05
#> math                     -2.218622e-04  1.824148e-05  5.369261e-06
#> progAcademic             -1.051474e-02 -1.025833e-03  6.643646e-06
#> progVocational           -5.911765e-03 -1.865578e-03 -1.019043e-04
#> (Intercept).precision     3.988849e-04 -4.684298e-04  8.386718e-06
#> gendermale.precision      7.371292e-04 -5.312241e-04 -3.678565e-05
#> math.precision            1.332653e-05 -3.548041e-05 -1.117223e-07
#> progAcademic.precision   -7.582598e-04  2.037610e-03  8.706788e-07
#> progVocational.precision -3.241746e-03  4.181426e-03  4.626451e-05
#>                           progAcademic progVocational (Intercept).precision
#> (Intercept)              -1.051474e-02  -5.911765e-03          3.988849e-04
#> gendermale               -1.025833e-03  -1.865578e-03         -4.684298e-04
#> math                      6.643646e-06  -1.019043e-04          8.386718e-06
#> progAcademic              1.746197e-02   1.062915e-02          2.197079e-04
#> progVocational            1.062915e-02   3.043066e-02         -3.347031e-03
#> (Intercept).precision     2.197079e-04  -3.347031e-03          1.178159e-01
#> gendermale.precision      7.818101e-04   3.612138e-03         -1.737732e-02
#> math.precision           -8.312163e-06   5.423776e-05         -7.945786e-04
#> progAcademic.precision   -2.333842e-04  -2.563226e-04         -7.751125e-02
#> progVocational.precision -2.082009e-04  -2.350747e-03         -6.340653e-02
#>                          gendermale.precision math.precision
#> (Intercept)                      7.371292e-04   1.332653e-05
#> gendermale                      -5.312241e-04  -3.548041e-05
#> math                            -3.678565e-05  -1.117223e-07
#> progAcademic                     7.818101e-04  -8.312163e-06
#> progVocational                   3.612138e-03   5.423776e-05
#> (Intercept).precision           -1.737732e-02  -7.945786e-04
#> gendermale.precision             4.149055e-02   4.782230e-05
#> math.precision                   4.782230e-05   1.748048e-05
#> progAcademic.precision          -3.021667e-03   5.731643e-05
#> progVocational.precision        -6.910452e-03  -2.345692e-04
#>                          progAcademic.precision progVocational.precision
#> (Intercept)                       -7.582598e-04            -3.241746e-03
#> gendermale                         2.037610e-03             4.181426e-03
#> math                               8.706788e-07             4.626451e-05
#> progAcademic                      -2.333842e-04            -2.082009e-04
#> progVocational                    -2.563226e-04            -2.350747e-03
#> (Intercept).precision             -7.751125e-02            -6.340653e-02
#> gendermale.precision              -3.021667e-03            -6.910452e-03
#> math.precision                     5.731643e-05            -2.345692e-04
#> progAcademic.precision             9.331766e-02             7.597928e-02
#> progVocational.precision           7.597928e-02             1.182064e-01

vcov(daysabs_fit, parameters = "mean")

#>                  (Intercept)    gendermale          math  progAcademic
#> (Intercept)     0.0217302395 -5.918840e-03 -2.218622e-04 -1.051474e-02
#> gendermale     -0.0059188402  1.391630e-02  1.824148e-05 -1.025833e-03
#> math           -0.0002218622  1.824148e-05  5.369261e-06  6.643646e-06
#> progAcademic   -0.0105147423 -1.025833e-03  6.643646e-06  1.746197e-02
#> progVocational -0.0059117646 -1.865578e-03 -1.019043e-04  1.062915e-02
#>                progVocational
#> (Intercept)     -0.0059117646
#> gendermale      -0.0018655785
#> math            -0.0001019043
#> progAcademic     0.0106291493
#> progVocational   0.0304306592

# }